Centering (Grand-Mean Centering)

Performs a grand-mean centering of data.

center(x, ...)

centre(x, ...)

# S3 method for numeric
center(x, weights = NULL, robust = FALSE, verbose = TRUE, ...)

# S3 method for data.frame
center(
  x,
  select = NULL,
  exclude = NULL,
  weights = NULL,
  robust = FALSE,
  force = FALSE,
  append = FALSE,
  suffix = "_c",
  verbose = TRUE,
  ...
)

Arguments

x
A data frame, a (numeric or character) vector or a factor.

...
Currently not used.

weights
Can be NULL (for no weighting), or:

For data frames: a numeric vector of weights, or a character of the name of a column in the data.frame that contains the weights.

For numeric vectors: a numeric vector of weights.

robust
Logical, if TRUE, centering is done by subtracting the median from the variables. If FALSE, variables are centered by subtracting the mean.

verbose
Toggle warnings and messages.

select
Character vector of column names. If NULL (the default), all variables will be selected.

exclude
Character vector of column names to be excluded from selection.

force
Logical, if TRUE, forces centering of factors as well. Factors are converted to numerical values, with the lowest level being the value 1 (unless the factor has numeric levels, which are converted to the corresponding numeric value).

append
Logical, if TRUE and x is a data frame, standardized variables will be added as additional columns; if FALSE, existing variables are overwritten.

suffix
Character value, will be appended to variable (column) names of x, if x is a data frame and append = TRUE.

Value

The centered variables.

Note

Difference between centering and standardizing: Standardized variables are computed by subtracting the mean of the variable and then dividing it by the standard deviation, while centering variables involves only the subtraction.

See also

Examples

data(iris)

# entire dataframe or a vector
head(iris$Sepal.Width)
#> [1] 3.5 3.0 3.2 3.1 3.6 3.9
head(center(iris$Sepal.Width))
#> [1]  0.44266667 -0.05733333  0.14266667  0.04266667  0.54266667  0.84266667
head(center(iris))
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1   -0.7433333  0.44266667       -2.358  -0.9993333  setosa
#> 2   -0.9433333 -0.05733333       -2.358  -0.9993333  setosa
#> 3   -1.1433333  0.14266667       -2.458  -0.9993333  setosa
#> 4   -1.2433333  0.04266667       -2.258  -0.9993333  setosa
#> 5   -0.8433333  0.54266667       -2.358  -0.9993333  setosa
#> 6   -0.4433333  0.84266667       -2.058  -0.7993333  setosa
head(center(iris, force = TRUE))
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1   -0.7433333  0.44266667       -2.358  -0.9993333      -1
#> 2   -0.9433333 -0.05733333       -2.358  -0.9993333      -1
#> 3   -1.1433333  0.14266667       -2.458  -0.9993333      -1
#> 4   -1.2433333  0.04266667       -2.258  -0.9993333      -1
#> 5   -0.8433333  0.54266667       -2.358  -0.9993333      -1
#> 6   -0.4433333  0.84266667       -2.058  -0.7993333      -1

# only the selected columns from a dataframe
center(anscombe, select = c("x1", "x3"))
#>    x1 x2 x3 x4    y1   y2    y3    y4
#> 1   1 10  1  8  8.04 9.14  7.46  6.58
#> 2  -1  8 -1  8  6.95 8.14  6.77  5.76
#> 3   4 13  4  8  7.58 8.74 12.74  7.71
#> 4   0  9  0  8  8.81 8.77  7.11  8.84
#> 5   2 11  2  8  8.33 9.26  7.81  8.47
#> 6   5 14  5  8  9.96 8.10  8.84  7.04
#> 7  -3  6 -3  8  7.24 6.13  6.08  5.25
#> 8  -5  4 -5 19  4.26 3.10  5.39 12.50
#> 9   3 12  3  8 10.84 9.13  8.15  5.56
#> 10 -2  7 -2  8  4.82 7.26  6.42  7.91
#> 11 -4  5 -4  8  5.68 4.74  5.73  6.89
center(anscombe, exclude = c("x1", "x3"))
#>    x1 x2 x3 x4          y1         y2    y3         y4
#> 1  10  1 10 -1  0.53909091  1.6390909 -0.04 -0.9209091
#> 2   8 -1  8 -1 -0.55090909  0.6390909 -0.73 -1.7409091
#> 3  13  4 13 -1  0.07909091  1.2390909  5.24  0.2090909
#> 4   9  0  9 -1  1.30909091  1.2690909 -0.39  1.3390909
#> 5  11  2 11 -1  0.82909091  1.7590909  0.31  0.9690909
#> 6  14  5 14 -1  2.45909091  0.5990909  1.34 -0.4609091
#> 7   6 -3  6 -1 -0.26090909 -1.3709091 -1.42 -2.2509091
#> 8   4 -5  4 10 -3.24090909 -4.4009091 -2.11  4.9990909
#> 9  12  3 12 -1  3.33909091  1.6290909  0.65 -1.9409091
#> 10  7 -2  7 -1 -2.68090909 -0.2409091 -1.08  0.4090909
#> 11  5 -4  5 -1 -1.82090909 -2.7609091 -1.77 -0.6109091