Extract a single column or element from an object

extract() is similar to $. It extracts a single column or element from an object (e.g., a data frame, list, )

data_extract(data, select, name = NULL, ...)

extract(data, select, name = NULL, ...)

Arguments

data

The object to subset. Methods are currently available for data frames and data frame extensions (e.g., tibbles).

select

A variable specified as:

a literal variable name (e.g., column_name)

a character vector with the variable name (e.g., "column_name")

a positive integer, giving the position counting from the left

a negative integer, giving the position counting from the right.

The default returns the last column. If the special value 0 or "row.names" is given, the row names of the object (if any) are extracted.

name

An optional argument that specifies the column to be used as names for for the vector after extraction. Specified in the same way as select.

...

For use by future methods.

Value

A vector containing the extracted element.

Examples

extract(mtcars, cyl, name = gear)
#> 4 4 4 3 3 3 3 4 4 4 4 3 3 3 3 3 3 4 4 4 3 3 3 3 3 4 5 5 5 5 5 4 
#> 6 6 4 6 8 6 8 4 4 6 6 8 8 8 8 8 8 4 4 4 4 8 8 8 8 4 4 4 8 6 8 4 
extract(mtcars, "cyl", name = gear)
#> 4 4 4 3 3 3 3 4 4 4 4 3 3 3 3 3 3 4 4 4 3 3 3 3 3 4 5 5 5 5 5 4 
#> 6 6 4 6 8 6 8 4 4 6 6 8 8 8 8 8 8 4 4 4 4 8 8 8 8 4 4 4 8 6 8 4 
extract(mtcars, -1, name = gear)
#> 4 4 4 3 3 3 3 4 4 4 4 3 3 3 3 3 3 4 4 4 3 3 3 3 3 4 5 5 5 5 5 4 
#> 4 4 1 1 2 1 4 2 2 4 4 3 3 3 4 4 4 1 2 1 1 2 2 4 2 1 2 2 4 6 8 2 
extract(mtcars, cyl, name = 0)
#>           Mazda RX4       Mazda RX4 Wag          Datsun 710      Hornet 4 Drive 
#>                   6                   6                   4                   6 
#>   Hornet Sportabout             Valiant          Duster 360           Merc 240D 
#>                   8                   6                   8                   4 
#>            Merc 230            Merc 280           Merc 280C          Merc 450SE 
#>                   4                   6                   6                   8 
#>          Merc 450SL         Merc 450SLC  Cadillac Fleetwood Lincoln Continental 
#>                   8                   8                   8                   8 
#>   Chrysler Imperial            Fiat 128         Honda Civic      Toyota Corolla 
#>                   8                   4                   4                   4 
#>       Toyota Corona    Dodge Challenger         AMC Javelin          Camaro Z28 
#>                   4                   8                   8                   8 
#>    Pontiac Firebird           Fiat X1-9       Porsche 914-2        Lotus Europa 
#>                   8                   4                   4                   4 
#>      Ford Pantera L        Ferrari Dino       Maserati Bora          Volvo 142E 
#>                   8                   6                   8                   4 
extract(mtcars, cyl, name = "row.names")
#>           Mazda RX4       Mazda RX4 Wag          Datsun 710      Hornet 4 Drive 
#>                   6                   6                   4                   6 
#>   Hornet Sportabout             Valiant          Duster 360           Merc 240D 
#>                   8                   6                   8                   4 
#>            Merc 230            Merc 280           Merc 280C          Merc 450SE 
#>                   4                   6                   6                   8 
#>          Merc 450SL         Merc 450SLC  Cadillac Fleetwood Lincoln Continental 
#>                   8                   8                   8                   8 
#>   Chrysler Imperial            Fiat 128         Honda Civic      Toyota Corolla 
#>                   8                   4                   4                   4 
#>       Toyota Corona    Dodge Challenger         AMC Javelin          Camaro Z28 
#>                   4                   8                   8                   8 
#>    Pontiac Firebird           Fiat X1-9       Porsche 914-2        Lotus Europa 
#>                   8                   4                   4                   4 
#>      Ford Pantera L        Ferrari Dino       Maserati Bora          Volvo 142E 
#>                   8                   6                   8                   4