Rescale Variables to a New Range

Rescale variables to a new range.

data_rescale(x, ...)

change_scale(x, ...)

# S3 method for numeric
data_rescale(x, to = c(0, 100), range = NULL, verbose = TRUE, ...)

# S3 method for grouped_df
data_rescale(
  x,
  to = c(0, 100),
  range = NULL,
  select = NULL,
  exclude = NULL,
  ...
)

# S3 method for data.frame
data_rescale(
  x,
  to = c(0, 100),
  range = NULL,
  select = NULL,
  exclude = NULL,
  ...
)

Arguments

x
A numeric variable.

...
Arguments passed to or from other methods.

to
New range that the variable will have after rescaling.

range
Initial (old) range of values. If NULL, will take the range of the input vector (range(x)).

verbose
Toggle warnings and messages on or off.

select
Character vector of column names. If NULL (the default), all variables will be selected.

exclude
Character vector of column names to be excluded from selection.

Value

A rescaled object.

See also

Examples

data_rescale(c(0, 1, 5, -5, -2))
#> [1]  50  60 100   0  30
data_rescale(c(0, 1, 5, -5, -2), to = c(-5, 5))
#> [1]  0  1  5 -5 -2

# Specify the "theoretical" range of the input vector
data_rescale(c(1, 3, 4), to = c(0, 40), range = c(0, 4))
#> [1] 10 30 40

# Dataframes
head(data_rescale(iris, to = c(0, 1)))
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1   0.22222222   0.6250000   0.06779661  0.04166667  setosa
#> 2   0.16666667   0.4166667   0.06779661  0.04166667  setosa
#> 3   0.11111111   0.5000000   0.05084746  0.04166667  setosa
#> 4   0.08333333   0.4583333   0.08474576  0.04166667  setosa
#> 5   0.19444444   0.6666667   0.06779661  0.04166667  setosa
#> 6   0.30555556   0.7916667   0.11864407  0.12500000  setosa
head(data_rescale(iris, to = c(0, 1), select = "Sepal.Length"))
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1   0.22222222         3.5          1.4         0.2  setosa
#> 2   0.16666667         3.0          1.4         0.2  setosa
#> 3   0.11111111         3.2          1.3         0.2  setosa
#> 4   0.08333333         3.1          1.5         0.2  setosa
#> 5   0.19444444         3.6          1.4         0.2  setosa
#> 6   0.30555556         3.9          1.7         0.4  setosa

# One can specify a list of ranges
head(data_rescale(iris, to = list("Sepal.Length" = c(0, 1),
                                  "Petal.Length" = c(-1, 0))))
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1   0.22222222         3.5   -0.9322034         0.2  setosa
#> 2   0.16666667         3.0   -0.9322034         0.2  setosa
#> 3   0.11111111         3.2   -0.9491525         0.2  setosa
#> 4   0.08333333         3.1   -0.9152542         0.2  setosa
#> 5   0.19444444         3.6   -0.9322034         0.2  setosa
#> 6   0.30555556         3.9   -0.8813559         0.4  setosa