data_restoretype {datawizard}	R Documentation

Restore the type of columns according to a reference data frame

Description

Restore the type of columns according to a reference data frame

Usage


data_restoretype(data, reference = NULL, ...)

Arguments

data
A data frame to pivot.

reference
A reference data frame from which to find the correct column types. If NULL, each column is converted to numeric if it doesn't generate NAs. For example, c("1", "2") can be converted to numeric but not c("Sepal.Length").

...
Currently not used.

Value

A data frame with columns whose types have been restored based on the reference data frame.

Examples


data <- data.frame(
  Sepal.Length = c("1", "3", "2"),
  Species = c("setosa", "versicolor", "setosa"),
  New = c("1", "3", "4")
)

fixed <- data_restoretype(data, reference = iris)
summary(fixed)

[Package datawizard version 0.10.0 Index]