Rotate a data frame

This function rotates a data frame, i.e. columns become rows and vice versa. It's the equivalent of using t() but restores the data.frame class, preserves attributes and prints a warning if the data type is modified (see example).

Usage

data_rotate(data, rownames = NULL, colnames = FALSE, verbose = TRUE)

Arguments

data

A data frame.

rownames

Character vector (optional). If not NULL, the data frame's rownames will be added as (first) column to the output, with rownames being the name of this column.

colnames

Logical (optional), if TRUE, the values of the first column in x will be used as column names in the rotated data frame.

verbose

Toggle warnings.

Value

A (rotated) data frame.

Examples

x <- mtcars[1:3, 1:4]
data_rotate(x)
#>      Mazda RX4 Mazda RX4 Wag Datsun 710
#> mpg         21            21       22.8
#> cyl          6             6        4.0
#> disp       160           160      108.0
#> hp         110           110       93.0
data_rotate(x, rownames = "property")
#>   property Mazda RX4 Mazda RX4 Wag Datsun 710
#> 1      mpg        21            21       22.8
#> 2      cyl         6             6        4.0
#> 3     disp       160           160      108.0
#> 4       hp       110           110       93.0

# use values in 1. column as column name
data_rotate(x, colnames = TRUE)
#>       21  21 22.8
#> cyl    6   6    4
#> disp 160 160  108
#> hp   110 110   93
data_rotate(x, rownames = "property", colnames = TRUE)
#>   property  21  21 22.8
#> 1      cyl   6   6    4
#> 2     disp 160 160  108
#> 3       hp 110 110   93

# warn that data types are changed
head(data_rotate(iris))
#> Warning: Your data frame contains mixed types of data. After transposition, all variables
#>   will be transformed into characters.
#>                   1      2      3      4      5      6      7      8      9
#> Sepal.Length    5.1    4.9    4.7    4.6    5.0    5.4    4.6    5.0    4.4
#> Sepal.Width     3.5    3.0    3.2    3.1    3.6    3.9    3.4    3.4    2.9
#> Petal.Length    1.4    1.4    1.3    1.5    1.4    1.7    1.4    1.5    1.4
#> Petal.Width     0.2    0.2    0.2    0.2    0.2    0.4    0.3    0.2    0.2
#> Species      setosa setosa setosa setosa setosa setosa setosa setosa setosa
#>                  10     11     12     13     14     15     16     17     18
#> Sepal.Length    4.9    5.4    4.8    4.8    4.3    5.8    5.7    5.4    5.1
#> Sepal.Width     3.1    3.7    3.4    3.0    3.0    4.0    4.4    3.9    3.5
#> Petal.Length    1.5    1.5    1.6    1.4    1.1    1.2    1.5    1.3    1.4
#> Petal.Width     0.1    0.2    0.2    0.1    0.1    0.2    0.4    0.4    0.3
#> Species      setosa setosa setosa setosa setosa setosa setosa setosa setosa
#>                  19     20     21     22     23     24     25     26     27
#> Sepal.Length    5.7    5.1    5.4    5.1    4.6    5.1    4.8    5.0    5.0
#> Sepal.Width     3.8    3.8    3.4    3.7    3.6    3.3    3.4    3.0    3.4
#> Petal.Length    1.7    1.5    1.7    1.5    1.0    1.7    1.9    1.6    1.6
#> Petal.Width     0.3    0.3    0.2    0.4    0.2    0.5    0.2    0.2    0.4
#> Species      setosa setosa setosa setosa setosa setosa setosa setosa setosa
#>                  28     29     30     31     32     33     34     35     36
#> Sepal.Length    5.2    5.2    4.7    4.8    5.4    5.2    5.5    4.9    5.0
#> Sepal.Width     3.5    3.4    3.2    3.1    3.4    4.1    4.2    3.1    3.2
#> Petal.Length    1.5    1.4    1.6    1.6    1.5    1.5    1.4    1.5    1.2
#> Petal.Width     0.2    0.2    0.2    0.2    0.4    0.1    0.2    0.2    0.2
#> Species      setosa setosa setosa setosa setosa setosa setosa setosa setosa
#>                  37     38     39     40     41     42     43     44     45
#> Sepal.Length    5.5    4.9    4.4    5.1    5.0    4.5    4.4    5.0    5.1
#> Sepal.Width     3.5    3.6    3.0    3.4    3.5    2.3    3.2    3.5    3.8
#> Petal.Length    1.3    1.4    1.3    1.5    1.3    1.3    1.3    1.6    1.9
#> Petal.Width     0.2    0.1    0.2    0.2    0.3    0.3    0.2    0.6    0.4
#> Species      setosa setosa setosa setosa setosa setosa setosa setosa setosa
#>                  46     47     48     49     50         51         52
#> Sepal.Length    4.8    5.1    4.6    5.3    5.0        7.0        6.4
#> Sepal.Width     3.0    3.8    3.2    3.7    3.3        3.2        3.2
#> Petal.Length    1.4    1.6    1.4    1.5    1.4        4.7        4.5
#> Petal.Width     0.3    0.2    0.2    0.2    0.2        1.4        1.5
#> Species      setosa setosa setosa setosa setosa versicolor versicolor
#>                      53         54         55         56         57         58
#> Sepal.Length        6.9        5.5        6.5        5.7        6.3        4.9
#> Sepal.Width         3.1        2.3        2.8        2.8        3.3        2.4
#> Petal.Length        4.9        4.0        4.6        4.5        4.7        3.3
#> Petal.Width         1.5        1.3        1.5        1.3        1.6        1.0
#> Species      versicolor versicolor versicolor versicolor versicolor versicolor
#>                      59         60         61         62         63         64
#> Sepal.Length        6.6        5.2        5.0        5.9        6.0        6.1
#> Sepal.Width         2.9        2.7        2.0        3.0        2.2        2.9
#> Petal.Length        4.6        3.9        3.5        4.2        4.0        4.7
#> Petal.Width         1.3        1.4        1.0        1.5        1.0        1.4
#> Species      versicolor versicolor versicolor versicolor versicolor versicolor
#>                      65         66         67         68         69         70
#> Sepal.Length        5.6        6.7        5.6        5.8        6.2        5.6
#> Sepal.Width         2.9        3.1        3.0        2.7        2.2        2.5
#> Petal.Length        3.6        4.4        4.5        4.1        4.5        3.9
#> Petal.Width         1.3        1.4        1.5        1.0        1.5        1.1
#> Species      versicolor versicolor versicolor versicolor versicolor versicolor
#>                      71         72         73         74         75         76
#> Sepal.Length        5.9        6.1        6.3        6.1        6.4        6.6
#> Sepal.Width         3.2        2.8        2.5        2.8        2.9        3.0
#> Petal.Length        4.8        4.0        4.9        4.7        4.3        4.4
#> Petal.Width         1.8        1.3        1.5        1.2        1.3        1.4
#> Species      versicolor versicolor versicolor versicolor versicolor versicolor
#>                      77         78         79         80         81         82
#> Sepal.Length        6.8        6.7        6.0        5.7        5.5        5.5
#> Sepal.Width         2.8        3.0        2.9        2.6        2.4        2.4
#> Petal.Length        4.8        5.0        4.5        3.5        3.8        3.7
#> Petal.Width         1.4        1.7        1.5        1.0        1.1        1.0
#> Species      versicolor versicolor versicolor versicolor versicolor versicolor
#>                      83         84         85         86         87         88
#> Sepal.Length        5.8        6.0        5.4        6.0        6.7        6.3
#> Sepal.Width         2.7        2.7        3.0        3.4        3.1        2.3
#> Petal.Length        3.9        5.1        4.5        4.5        4.7        4.4
#> Petal.Width         1.2        1.6        1.5        1.6        1.5        1.3
#> Species      versicolor versicolor versicolor versicolor versicolor versicolor
#>                      89         90         91         92         93         94
#> Sepal.Length        5.6        5.5        5.5        6.1        5.8        5.0
#> Sepal.Width         3.0        2.5        2.6        3.0        2.6        2.3
#> Petal.Length        4.1        4.0        4.4        4.6        4.0        3.3
#> Petal.Width         1.3        1.3        1.2        1.4        1.2        1.0
#> Species      versicolor versicolor versicolor versicolor versicolor versicolor
#>                      95         96         97         98         99        100
#> Sepal.Length        5.6        5.7        5.7        6.2        5.1        5.7
#> Sepal.Width         2.7        3.0        2.9        2.9        2.5        2.8
#> Petal.Length        4.2        4.2        4.2        4.3        3.0        4.1
#> Petal.Width         1.3        1.2        1.3        1.3        1.1        1.3
#> Species      versicolor versicolor versicolor versicolor versicolor versicolor
#>                    101       102       103       104       105       106
#> Sepal.Length       6.3       5.8       7.1       6.3       6.5       7.6
#> Sepal.Width        3.3       2.7       3.0       2.9       3.0       3.0
#> Petal.Length       6.0       5.1       5.9       5.6       5.8       6.6
#> Petal.Width        2.5       1.9       2.1       1.8       2.2       2.1
#> Species      virginica virginica virginica virginica virginica virginica
#>                    107       108       109       110       111       112
#> Sepal.Length       4.9       7.3       6.7       7.2       6.5       6.4
#> Sepal.Width        2.5       2.9       2.5       3.6       3.2       2.7
#> Petal.Length       4.5       6.3       5.8       6.1       5.1       5.3
#> Petal.Width        1.7       1.8       1.8       2.5       2.0       1.9
#> Species      virginica virginica virginica virginica virginica virginica
#>                    113       114       115       116       117       118
#> Sepal.Length       6.8       5.7       5.8       6.4       6.5       7.7
#> Sepal.Width        3.0       2.5       2.8       3.2       3.0       3.8
#> Petal.Length       5.5       5.0       5.1       5.3       5.5       6.7
#> Petal.Width        2.1       2.0       2.4       2.3       1.8       2.2
#> Species      virginica virginica virginica virginica virginica virginica
#>                    119       120       121       122       123       124
#> Sepal.Length       7.7       6.0       6.9       5.6       7.7       6.3
#> Sepal.Width        2.6       2.2       3.2       2.8       2.8       2.7
#> Petal.Length       6.9       5.0       5.7       4.9       6.7       4.9
#> Petal.Width        2.3       1.5       2.3       2.0       2.0       1.8
#> Species      virginica virginica virginica virginica virginica virginica
#>                    125       126       127       128       129       130
#> Sepal.Length       6.7       7.2       6.2       6.1       6.4       7.2
#> Sepal.Width        3.3       3.2       2.8       3.0       2.8       3.0
#> Petal.Length       5.7       6.0       4.8       4.9       5.6       5.8
#> Petal.Width        2.1       1.8       1.8       1.8       2.1       1.6
#> Species      virginica virginica virginica virginica virginica virginica
#>                    131       132       133       134       135       136
#> Sepal.Length       7.4       7.9       6.4       6.3       6.1       7.7
#> Sepal.Width        2.8       3.8       2.8       2.8       2.6       3.0
#> Petal.Length       6.1       6.4       5.6       5.1       5.6       6.1
#> Petal.Width        1.9       2.0       2.2       1.5       1.4       2.3
#> Species      virginica virginica virginica virginica virginica virginica
#>                    137       138       139       140       141       142
#> Sepal.Length       6.3       6.4       6.0       6.9       6.7       6.9
#> Sepal.Width        3.4       3.1       3.0       3.1       3.1       3.1
#> Petal.Length       5.6       5.5       4.8       5.4       5.6       5.1
#> Petal.Width        2.4       1.8       1.8       2.1       2.4       2.3
#> Species      virginica virginica virginica virginica virginica virginica
#>                    143       144       145       146       147       148
#> Sepal.Length       5.8       6.8       6.7       6.7       6.3       6.5
#> Sepal.Width        2.7       3.2       3.3       3.0       2.5       3.0
#> Petal.Length       5.1       5.9       5.7       5.2       5.0       5.2
#> Petal.Width        1.9       2.3       2.5       2.3       1.9       2.0
#> Species      virginica virginica virginica virginica virginica virginica
#>                    149       150
#> Sepal.Length       6.2       5.9
#> Sepal.Width        3.4       3.0
#> Petal.Length       5.4       5.1
#> Petal.Width        2.3       1.8
#> Species      virginica virginica