Create frequency tables of variables

This function creates frequency tables of variables, including the number of levels/values as well as the distribution of raw, valid and cumulative percentages.

Usage

data_tabulate(x, ...)

# S3 method for default
data_tabulate(x, drop_levels = FALSE, name = NULL, verbose = TRUE, ...)

# S3 method for data.frame
data_tabulate(
  x,
  select = NULL,
  exclude = NULL,
  ignore_case = FALSE,
  regex = FALSE,
  collapse = FALSE,
  drop_levels = FALSE,
  verbose = TRUE,
  ...
)

Arguments

x

A (grouped) data frame, a vector or factor.

...

not used.

drop_levels

Logical, if TRUE, factor levels that do not occur in the data are included in the table (with frequency of zero), else unused factor levels are dropped from the frequency table.

name

Optional character string, which includes the name that is used for printing.

verbose

Toggle warnings.

select

Variables that will be included when performing the required tasks. Can be either

a variable specified as a literal variable name (e.g., column_name),

a string with the variable name (e.g., "column_name"), or a character vector of variable names (e.g., c("col1", "col2", "col3")),

a formula with variable names (e.g., ~column_1 + column_2),

a vector of positive integers, giving the positions counting from the left (e.g. 1 or c(1, 3, 5)),

a vector of negative integers, giving the positions counting from the right (e.g., -1 or -1:-3),

one of the following select-helpers: starts_with(), ends_with(), contains(), a range using : or regex(""). starts_with(), ends_with(), and contains() accept several patterns, e.g starts_with("Sep", "Petal").

or a function testing for logical conditions, e.g. is.numeric() (or is.numeric), or any user-defined function that selects the variables for which the function returns TRUE (like: foo <- function(x) mean(x) > 3),

ranges specified via literal variable names, select-helpers (except regex()) and (user-defined) functions can be negated, i.e. return non-matching elements, when prefixed with a -, e.g. -ends_with(""), -is.numeric or -Sepal.Width:Petal.Length. Note: Negation means that matches are excluded, and thus, the exclude argument can be used alternatively. For instance, select=-ends_with("Length") (with -) is equivalent to exclude=ends_with("Length") (no -). In case negation should not work as expected, use the exclude argument instead.

If NULL, selects all columns. Patterns that found no matches are silently ignored, e.g. find_columns(iris, select = c("Species", "Test")) will just return "Species".

exclude

See select, however, column names matched by the pattern from exclude will be excluded instead of selected. If NULL (the default), excludes no columns.

ignore_case

Logical, if TRUE and when one of the select-helpers or a regular expression is used in select, ignores lower/upper case in the search pattern when matching against variable names.

regex

Logical, if TRUE, the search pattern from select will be treated as regular expression. When regex = TRUE, select must be a character string (or a variable containing a character string) and is not allowed to be one of the supported select-helpers or a character vector of length > 1. regex = TRUE is comparable to using one of the two select-helpers, select = contains("") or select = regex(""), however, since the select-helpers may not work when called from inside other functions (see 'Details'), this argument may be used as workaround.

collapse

Logical, if TRUE collapses multiple tables into one larger table for printing. This affects only printing, not the returned object.

Value

A data frame, or a list of data frames, with one frequency table as data frame per variable.

Examples

data(efc)

# vector/factor
data_tabulate(efc$c172code)
#> carer's level of education (efc$c172code) <numeric>
#> # total N=100 valid N=90
#> 
#> Value |  N | Raw % | Valid % | Cumulative %
#> ------+----+-------+---------+-------------
#> 1     |  8 |  8.00 |    8.89 |         8.89
#> 2     | 66 | 66.00 |   73.33 |        82.22
#> 3     | 16 | 16.00 |   17.78 |       100.00
#> <NA>  | 10 | 10.00 |    <NA> |         <NA>

# data frame
data_tabulate(efc, c("e42dep", "c172code"))
#> elder's dependency (e42dep) <categorical>
#> # total N=100 valid N=97
#> 
#> Value |  N | Raw % | Valid % | Cumulative %
#> ------+----+-------+---------+-------------
#> 1     |  2 |  2.00 |    2.06 |         2.06
#> 2     |  4 |  4.00 |    4.12 |         6.19
#> 3     | 28 | 28.00 |   28.87 |        35.05
#> 4     | 63 | 63.00 |   64.95 |       100.00
#> <NA>  |  3 |  3.00 |    <NA> |         <NA>
#> 
#> carer's level of education (c172code) <numeric>
#> # total N=100 valid N=90
#> 
#> Value |  N | Raw % | Valid % | Cumulative %
#> ------+----+-------+---------+-------------
#> 1     |  8 |  8.00 |    8.89 |         8.89
#> 2     | 66 | 66.00 |   73.33 |        82.22
#> 3     | 16 | 16.00 |   17.78 |       100.00
#> <NA>  | 10 | 10.00 |    <NA> |         <NA>

# grouped data frame
suppressPackageStartupMessages(library(poorman, quietly = TRUE))
efc %>%
  group_by(c172code) %>%
  data_tabulate("e16sex")
#> elder's gender (e16sex) <numeric>
#> Grouped by c172code (1)
#> # total N=8 valid N=8
#> 
#> Value | N | Raw % | Valid % | Cumulative %
#> ------+---+-------+---------+-------------
#> 1     | 5 | 62.50 |   62.50 |        62.50
#> 2     | 3 | 37.50 |   37.50 |       100.00
#> <NA>  | 0 |  0.00 |    <NA> |         <NA>
#> 
#> elder's gender (e16sex) <numeric>
#> Grouped by c172code (2)
#> # total N=66 valid N=66
#> 
#> Value |  N | Raw % | Valid % | Cumulative %
#> ------+----+-------+---------+-------------
#> 1     | 32 | 48.48 |   48.48 |        48.48
#> 2     | 34 | 51.52 |   51.52 |       100.00
#> <NA>  |  0 |  0.00 |    <NA> |         <NA>
#> 
#> elder's gender (e16sex) <numeric>
#> Grouped by c172code (3)
#> # total N=16 valid N=16
#> 
#> Value |  N | Raw % | Valid % | Cumulative %
#> ------+----+-------+---------+-------------
#> 1     |  4 | 25.00 |   25.00 |        25.00
#> 2     | 12 | 75.00 |   75.00 |       100.00
#> <NA>  |  0 |  0.00 |    <NA> |         <NA>
#> 
#> elder's gender (e16sex) <numeric>
#> Grouped by c172code (NA)
#> # total N=10 valid N=10
#> 
#> Value | N | Raw % | Valid % | Cumulative %
#> ------+---+-------+---------+-------------
#> 1     | 5 | 50.00 |   50.00 |        50.00
#> 2     | 5 | 50.00 |   50.00 |       100.00
#> <NA>  | 0 |  0.00 |    <NA> |         <NA>

# collapse tables
efc %>%
  group_by(c172code) %>%
  data_tabulate("e16sex", collapse = TRUE)
#> # Frequency Table
#> 
#> Variable |         Group | Value |  N | Raw % | Valid % | Cumulative %
#> ---------+---------------+-------+----+-------+---------+-------------
#> e16sex   |  c172code (1) |     1 |  5 | 62.50 |   62.50 |        62.50
#>          |               |     2 |  3 | 37.50 |   37.50 |       100.00
#>          |               |  <NA> |  0 |  0.00 |    <NA> |         <NA>
#> ---------+---------------+-------+----+-------+---------+-------------
#> e16sex   |  c172code (2) |     1 | 32 | 48.48 |   48.48 |        48.48
#>          |               |     2 | 34 | 51.52 |   51.52 |       100.00
#>          |               |  <NA> |  0 |  0.00 |    <NA> |         <NA>
#> ---------+---------------+-------+----+-------+---------+-------------
#> e16sex   |  c172code (3) |     1 |  4 | 25.00 |   25.00 |        25.00
#>          |               |     2 | 12 | 75.00 |   75.00 |       100.00
#>          |               |  <NA> |  0 |  0.00 |    <NA> |         <NA>
#> ---------+---------------+-------+----+-------+---------+-------------
#> e16sex   | c172code (NA) |     1 |  5 | 50.00 |   50.00 |        50.00
#>          |               |     2 |  5 | 50.00 |   50.00 |       100.00
#>          |               |  <NA> |  0 |  0.00 |    <NA> |         <NA>
#> ----------------------------------------------------------------------

# for larger N's (> 100000), a big mark is automatically added
set.seed(123)
x <- sample(1:3, 1e6, TRUE)
data_tabulate(x, name = "Large Number")
#> Large Number (x) <integer>
#> # total N=1000000 valid N=1000000
#> 
#> Value |       N | Raw % | Valid % | Cumulative %
#> ------+---------+-------+---------+-------------
#> 1     | 333,852 | 33.39 |   33.39 |        33.39
#> 2     | 332,910 | 33.29 |   33.29 |        66.68
#> 3     | 333,238 | 33.32 |   33.32 |       100.00
#> <NA>  |       0 |  0.00 |    <NA> |         <NA>

# to remove the big mark, use "print(..., big_mark = "")"
print(data_tabulate(x), big_mark = "")
#> x <integer>
#> # total N=1000000 valid N=1000000
#> 
#> Value |      N | Raw % | Valid % | Cumulative %
#> ------+--------+-------+---------+-------------
#> 1     | 333852 | 33.39 |   33.39 |        33.39
#> 2     | 332910 | 33.29 |   33.29 |        66.68
#> 3     | 333238 | 33.32 |   33.32 |       100.00
#> <NA>  |      0 |  0.00 |    <NA> |         <NA>