datawizard-package {datawizard}	R Documentation

datawizard: Easy Data Wrangling and Statistical Transformations

Description

A lightweight package to assist in key steps involved in any data analysis workflow:

wrangling the raw data to get it in the needed form,

applying preprocessing steps and statistical transformations, and

compute statistical summaries of data properties and distributions.

It is also the data wrangling backend for packages in 'easystats' ecosystem. References: Patil et al. (2022) doi: 10.21105/joss.04684.

Details

datawizard

Author(s)

Maintainer: Etienne Bacher etienne.bacher@protonmail.com (ORCID)

Authors:

Indrajeet Patil patilindrajeet.science@gmail.com (ORCID) (@patilindrajeets)

Dominique Makowski dom.makowski@gmail.com (ORCID) (@Dom_Makowski)

Daniel Lüdecke d.luedecke@uke.de (ORCID) (@strengejacke)

Mattan S. Ben-Shachar matanshm@post.bgu.ac.il (ORCID)

Brenton M. Wiernik brenton@wiernik.org (ORCID) (@bmwiernik)

Other contributors:

Rémi Thériault remi.theriault@mail.mcgill.ca (ORCID) (@rempsyc) [contributor]

Thomas J. Faulkenberry faulkenberry@tarleton.edu [reviewer]

Robert Garrett rcg4@illinois.edu [reviewer]

See Also

Useful links:

https://easystats.github.io/datawizard/

Report bugs at https://github.com/easystats/datawizard/issues

[Package datawizard version 0.10.0 Index]