.is_deprecated {datawizard}	R Documentation

Print a message saying that an argument is deprecated and that the user should use its replacement instead.

Description

Print a message saying that an argument is deprecated and that the user should use its replacement instead.

Usage


.is_deprecated(arg, replacement)

Arguments

arg
Argument that is deprecated

replacement
Argument that replaces the deprecated argument

[Package datawizard version 0.10.0 Index]