Normalize numeric variable to 0-1 range

Performs a normalization of data, i.e., it scales variables in the range 0 - 1. This is a special case of rescale(). unnormalize() is the counterpart, but only works for variables that have been normalized with normalize().

normalize(x, ...)

# S3 method for numeric
normalize(x, include_bounds = TRUE, verbose = TRUE, ...)

# S3 method for data.frame
normalize(
  x,
  select = NULL,
  exclude = NULL,
  include_bounds = TRUE,
  ignore_case = FALSE,
  verbose = TRUE,
  ...
)

unnormalize(x, ...)

# S3 method for numeric
unnormalize(x, verbose = TRUE, ...)

# S3 method for data.frame
unnormalize(
  x,
  select = NULL,
  exclude = NULL,
  ignore_case = FALSE,
  verbose = TRUE,
  ...
)

Arguments

x

A numeric vector, (grouped) data frame, or matrix. See 'Details'.

...

Arguments passed to or from other methods.

include_bounds

Logical, if TRUE, return value may include 0 and 1. If FALSE, the return value is compressed, using Smithson and Verkuilen's (2006) formula (x * (n - 1) + 0.5) / n, to avoid zeros and ones in the normalized variables. This can be useful in case of beta-regression, where the response variable is not allowed to include zeros and ones.

verbose

Toggle warnings and messages on or off.

select

Variables that will be included when performing the required tasks. Can be either

a variable specified as a literal variable name (e.g., column_name),

a string with the variable name (e.g., "column_name"), or a character vector of variable names (e.g., c("col1", "col2", "col3")),

a formula with variable names (e.g., ~column_1 + column_2),

a vector of positive integers, giving the positions counting from the left (e.g. 1 or c(1, 3, 5)),

a vector of negative integers, giving the positions counting from the right (e.g., -1 or -1:-3),

one of the following select-helpers: starts_with(""), ends_with(""), contains(""), a range using : or regex(""),

or a function testing for logical conditions, e.g. is.numeric() (or is.numeric), or any user-defined function that selects the variables for which the function returns TRUE (like: foo <- function(x) mean(x) > 3),

ranges specified via literal variable names, select-helpers (except regex()) and (user-defined) functions can be negated, i.e. return non-matching elements, when prefixed with a -, e.g. -ends_with(""), -is.numeric or -Sepal.Width:Petal.Length. Note: Negation means that matches are excluded, and thus, the exclude argument can be used alternatively. For instance, select=-ends_with("Length") (with -) is equivalent to exclude=ends_with("Length") (no -). In case negation should not work as expected, use the exclude argument instead.

If NULL, selects all columns. Patterns that found no matches are silently ignored, e.g. find_columns(iris, select = c("Species", "Test")) will just return "Species".

exclude

See select, however, column names matched by the pattern from exclude will be excluded instead of selected. If NULL (the default), excludes no columns.

ignore_case

Logical, if TRUE and when one of the select-helpers or a regular expression is used in select, ignores lower/upper case in the search pattern when matching against variable names.

Value

A normalized object.

Details

If x is a matrix, normalization is performed across all values (not column- or row-wise). For column-wise normalization, convert the matrix to a data.frame.

If x is a grouped data frame (grouped_df), normalization is performed separately for each group.

Selection of variables - the select argument

For most functions that have a select argument (including this function), the complete input data frame is returned, even when select only selects a range of variables. That is, the function is only applied to those variables that have a match in select, while all other variables remain unchanged. In other words: for this function, select will not omit any non-included variables, so that the returned data frame will include all variables from the input data frame.

References

Smithson M, Verkuilen J (2006). A Better Lemon Squeezer? Maximum-Likelihood Regression with Beta-Distributed Dependent Variables. Psychological Methods, 11(1), 54–71.

See also

Examples

normalize(c(0, 1, 5, -5, -2))
#> [1] 0.5 0.6 1.0 0.0 0.3
#> attr(,"include_bounds")
#> [1] TRUE
#> attr(,"min_value")
#> [1] -5
#> attr(,"range_difference")
#> [1] 10
normalize(c(0, 1, 5, -5, -2), include_bounds = FALSE)
#> [1] 0.50 0.58 0.90 0.10 0.34
#> attr(,"include_bounds")
#> [1] FALSE
#> attr(,"min_value")
#> [1] -5
#> attr(,"range_difference")
#> [1] 10

head(normalize(trees))
#>        Girth     Height      Volume
#> 1 0.00000000 0.29166667 0.001497006
#> 2 0.02439024 0.08333333 0.001497006
#> 3 0.04065041 0.00000000 0.000000000
#> 4 0.17886179 0.37500000 0.092814371
#> 5 0.19512195 0.75000000 0.128742515
#> 6 0.20325203 0.83333333 0.142215569