Convenient dataframe manipulation functionalities

Safe and intuitive functions to manipulate dataframes.

data_addprefix(data, pattern, ...)

data_addsuffix(data, pattern, ...)

data_findcols(data, pattern = NULL, starts_with = NULL, ends_with = NULL, ...)

data_remove(data, pattern, ...)

data_rename(data, pattern = NULL, replacement = NULL, safe = TRUE, ...)

data_rename_rows(data, rows = NULL)

data_reorder(data, cols, safe = TRUE, ...)

Arguments

data

A data frame, or an object that can be coerced to a data frame.

pattern, replacement, starts_with, ends_with

Character strings.

...

Other arguments passed to or from other functions.

safe

Do not throw error if for instance the variable to be renamed/removed doesn't exist.

cols, rows

Vector of column or row names.

Value

A modified data frame.

Examples

# Add prefix / suffix to all columns
head(data_addprefix(iris, "NEW_"))
#>   NEW_Sepal.Length NEW_Sepal.Width NEW_Petal.Length NEW_Petal.Width NEW_Species
#> 1              5.1             3.5              1.4             0.2      setosa
#> 2              4.9             3.0              1.4             0.2      setosa
#> 3              4.7             3.2              1.3             0.2      setosa
#> 4              4.6             3.1              1.5             0.2      setosa
#> 5              5.0             3.6              1.4             0.2      setosa
#> 6              5.4             3.9              1.7             0.4      setosa
head(data_addsuffix(iris, "_OLD"))
#>   Sepal.Length_OLD Sepal.Width_OLD Petal.Length_OLD Petal.Width_OLD Species_OLD
#> 1              5.1             3.5              1.4             0.2      setosa
#> 2              4.9             3.0              1.4             0.2      setosa
#> 3              4.7             3.2              1.3             0.2      setosa
#> 4              4.6             3.1              1.5             0.2      setosa
#> 5              5.0             3.6              1.4             0.2      setosa
#> 6              5.4             3.9              1.7             0.4      setosa
# Find columns names by pattern
data_findcols(iris, starts_with = "Sepal")
#> [1] "Sepal.Length" "Sepal.Width" 
data_findcols(iris, ends_with = "Width")
#> [1] "Sepal.Width" "Petal.Width"
data_findcols(iris, pattern = "\\.")
#> [1] "Sepal.Length" "Sepal.Width"  "Petal.Length" "Petal.Width" 
data_findcols(iris, c("Petal.Width", "Sepal.Length"))
#> [1] "Petal.Width"  "Sepal.Length"
# Remove columns
head(data_remove(iris, "Sepal.Length"))
#>   Sepal.Width Petal.Length Petal.Width Species
#> 1         3.5          1.4         0.2  setosa
#> 2         3.0          1.4         0.2  setosa
#> 3         3.2          1.3         0.2  setosa
#> 4         3.1          1.5         0.2  setosa
#> 5         3.6          1.4         0.2  setosa
#> 6         3.9          1.7         0.4  setosa
# Rename columns
head(data_rename(iris, "Sepal.Length", "length"))
#>   length Sepal.Width Petal.Length Petal.Width Species
#> 1    5.1         3.5          1.4         0.2  setosa
#> 2    4.9         3.0          1.4         0.2  setosa
#> 3    4.7         3.2          1.3         0.2  setosa
#> 4    4.6         3.1          1.5         0.2  setosa
#> 5    5.0         3.6          1.4         0.2  setosa
#> 6    5.4         3.9          1.7         0.4  setosa
# data_rename(iris, "FakeCol", "length", safe=FALSE)  # This fails
head(data_rename(iris, "FakeCol", "length")) # This doesn't
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa
#> 4          4.6         3.1          1.5         0.2  setosa
#> 5          5.0         3.6          1.4         0.2  setosa
#> 6          5.4         3.9          1.7         0.4  setosa
head(data_rename(iris, c("Sepal.Length", "Sepal.Width"), c("length", "width")))
#>   length width Petal.Length Petal.Width Species
#> 1    5.1   3.5          1.4         0.2  setosa
#> 2    4.9   3.0          1.4         0.2  setosa
#> 3    4.7   3.2          1.3         0.2  setosa
#> 4    4.6   3.1          1.5         0.2  setosa
#> 5    5.0   3.6          1.4         0.2  setosa
#> 6    5.4   3.9          1.7         0.4  setosa

# Reset names
head(data_rename(iris, NULL))
#>     1   2   3   4      5
#> 1 5.1 3.5 1.4 0.2 setosa
#> 2 4.9 3.0 1.4 0.2 setosa
#> 3 4.7 3.2 1.3 0.2 setosa
#> 4 4.6 3.1 1.5 0.2 setosa
#> 5 5.0 3.6 1.4 0.2 setosa
#> 6 5.4 3.9 1.7 0.4 setosa

# Change all
head(data_rename(iris, paste0("Var", 1:5)))
#>   Var1 Var2 Var3 Var4   Var5
#> 1  5.1  3.5  1.4  0.2 setosa
#> 2  4.9  3.0  1.4  0.2 setosa
#> 3  4.7  3.2  1.3  0.2 setosa
#> 4  4.6  3.1  1.5  0.2 setosa
#> 5  5.0  3.6  1.4  0.2 setosa
#> 6  5.4  3.9  1.7  0.4 setosa
# Reorder columns
head(data_reorder(iris, c("Species", "Sepal.Length")))
#>   Species Sepal.Length Sepal.Width Petal.Length Petal.Width
#> 1  setosa          5.1         3.5          1.4         0.2
#> 2  setosa          4.9         3.0          1.4         0.2
#> 3  setosa          4.7         3.2          1.3         0.2
#> 4  setosa          4.6         3.1          1.5         0.2
#> 5  setosa          5.0         3.6          1.4         0.2
#> 6  setosa          5.4         3.9          1.7         0.4
head(data_reorder(iris, c("Species", "dupa"))) # Safe for non-existing cols
#>   Species Sepal.Length Sepal.Width Petal.Length Petal.Width
#> 1  setosa          5.1         3.5          1.4         0.2
#> 2  setosa          4.9         3.0          1.4         0.2
#> 3  setosa          4.7         3.2          1.3         0.2
#> 4  setosa          4.6         3.1          1.5         0.2
#> 5  setosa          5.0         3.6          1.4         0.2
#> 6  setosa          5.4         3.9          1.7         0.4