Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data
A data frame to pivot.

cols
A vector of column names or indices to pivot into longer format.

colnames_to
The name of the new column that will contain the column names.

values_to
The name of the new column that will contain the values of the pivoted variables.

rows_to
The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...
Additional arguments passed on to methods.

names_to, names_from
Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from
The name of the column that contains the values of the put in the columns.

colnames_from
The name of the column that contains the levels to be used as future columns.

rows_from
The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep
The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name        Value
#> 1    X1  0.403741445
#> 2    X2  2.522827287
#> 3    X3 -0.384572915
#> 4    X4  0.140054453
#> 5    X5 -0.067004301
#> 6    X1  0.065058897
#> 7    X2 -0.775335916
#> 8    X3  0.904951044
#> 9    X4  1.122344129
#> 10   X5  0.051716301
#> 11   X1 -0.800352348
#> 12   X2 -0.658709001
#> 13   X3 -0.547465764
#> 14   X4 -0.978016568
#> 15   X5 -1.297680017
#> 16   X1 -0.221141102
#> 17   X2 -0.416464924
#> 18   X3 -0.109990634
#> 19   X4  0.388207088
#> 20   X5  0.754107106
#> 21   X1 -0.741640017
#> 22   X2 -0.008003526
#> 23   X3 -0.199626570
#> 24   X4 -0.137828552
#> 25   X5  1.170098976
#> 26   X1  0.242260939
#> 27   X2 -0.687387110
#> 28   X3  0.003486649
#> 29   X4 -1.370036661
#> 30   X5 -1.092270314
#> 31   X1 -0.348637778
#> 32   X2 -0.722534948
#> 33   X3 -0.238590803
#> 34   X4 -2.051494125
#> 35   X5  0.242553639
#> 36   X1 -2.395427880
#> 37   X2  1.035658325
#> 38   X3  1.973022190
#> 39   X4 -0.731621347
#> 40   X5  1.200928290
#> 41   X1  0.192347182
#> 42   X2  1.111670161
#> 43   X3  0.229472860
#> 44   X4  0.072370811
#> 45   X5  0.445897459
#> 46   X1  0.719988944
#> 47   X2 -0.396322899
#> 48   X3  0.289430087
#> 49   X4  1.145167580
#> 50   X5  0.117286222

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>              X3          X4         X5 Row Column      Numbers
#> 1  -0.384572915  0.14005445 -0.0670043   1     X1  0.403741445
#> 2  -0.384572915  0.14005445 -0.0670043   1     X2  2.522827287
#> 3   0.904951044  1.12234413  0.0517163   2     X1  0.065058897
#> 4   0.904951044  1.12234413  0.0517163   2     X2 -0.775335916
#> 5  -0.547465764 -0.97801657 -1.2976800   3     X1 -0.800352348
#> 6  -0.547465764 -0.97801657 -1.2976800   3     X2 -0.658709001
#> 7  -0.109990634  0.38820709  0.7541071   4     X1 -0.221141102
#> 8  -0.109990634  0.38820709  0.7541071   4     X2 -0.416464924
#> 9  -0.199626570 -0.13782855  1.1700990   5     X1 -0.741640017
#> 10 -0.199626570 -0.13782855  1.1700990   5     X2 -0.008003526
#> 11  0.003486649 -1.37003666 -1.0922703   6     X1  0.242260939
#> 12  0.003486649 -1.37003666 -1.0922703   6     X2 -0.687387110
#> 13 -0.238590803 -2.05149413  0.2425536   7     X1 -0.348637778
#> 14 -0.238590803 -2.05149413  0.2425536   7     X2 -0.722534948
#> 15  1.973022190 -0.73162135  1.2009283   8     X1 -2.395427880
#> 16  1.973022190 -0.73162135  1.2009283   8     X2  1.035658325
#> 17  0.229472860  0.07237081  0.4458975   9     X1  0.192347182
#> 18  0.229472860  0.07237081  0.4458975   9     X2  1.111670161
#> 19  0.289430087  1.14516758  0.1172862  10     X1  0.719988944
#> 20  0.289430087  1.14516758  0.1172862  10     X2 -0.396322899

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID   Value_X1     Value_X2     Value_X3    Value_X4   Value_X5
#> 1       1  0.4037414  2.522827287 -0.384572915  0.14005445 -0.0670043
#> 2       2  0.0650589 -0.775335916  0.904951044  1.12234413  0.0517163
#> 3       3 -0.8003523 -0.658709001 -0.547465764 -0.97801657 -1.2976800
#> 4       4 -0.2211411 -0.416464924 -0.109990634  0.38820709  0.7541071
#> 5       5 -0.7416400 -0.008003526 -0.199626570 -0.13782855  1.1700990
#> 6       6  0.2422609 -0.687387110  0.003486649 -1.37003666 -1.0922703
#> 7       7 -0.3486378 -0.722534948 -0.238590803 -2.05149413  0.2425536
#> 8       8 -2.3954279  1.035658325  1.973022190 -0.73162135  1.2009283
#> 9       9  0.1923472  1.111670161  0.229472860  0.07237081  0.4458975
#> 10     10  0.7199889 -0.396322899  0.289430087  1.14516758  0.1172862

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5