Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns.

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name        Value
#> 1    X1 -0.531005363
#> 2    X2 -0.253775448
#> 3    X3 -1.444357328
#> 4    X4 -0.655343538
#> 5    X5  0.435993599
#> 6    X1  0.809429932
#> 7    X2 -1.034383520
#> 8    X3 -1.428263620
#> 9    X4 -0.280563081
#> 10   X5  0.589693932
#> 11   X1  1.753244629
#> 12   X2  1.277862048
#> 13   X3 -1.196961925
#> 14   X4 -0.464708181
#> 15   X5  1.738934532
#> 16   X1  0.340333778
#> 17   X2 -0.493152888
#> 18   X3 -0.123501536
#> 19   X4 -0.153740777
#> 20   X5 -0.474367129
#> 21   X1  0.057425469
#> 22   X2 -0.150566577
#> 23   X3 -0.326619288
#> 24   X4  0.440330269
#> 25   X5  1.063352748
#> 26   X1 -1.072287276
#> 27   X2  0.304600857
#> 28   X3  0.048358448
#> 29   X4 -0.819096837
#> 30   X5 -0.046273080
#> 31   X1 -1.393680620
#> 32   X2  0.725087948
#> 33   X3 -0.278517765
#> 34   X4 -0.012241473
#> 35   X5  0.598518174
#> 36   X1 -0.478111108
#> 37   X2 -1.422826562
#> 38   X3  0.175708534
#> 39   X4 -0.009096347
#> 40   X5 -1.783898352
#> 41   X1 -0.197234780
#> 42   X2 -1.788009161
#> 43   X3  0.877338281
#> 44   X4 -0.644358171
#> 45   X5 -0.690714857
#> 46   X1 -0.956997309
#> 47   X2  1.850277846
#> 48   X3  1.160332311
#> 49   X4 -1.832799747
#> 50   X5 -1.367560163

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>             X3           X4          X5 Row Column     Numbers
#> 1  -1.44435733 -0.655343538  0.43599360   1     X1 -0.53100536
#> 2  -1.44435733 -0.655343538  0.43599360   1     X2 -0.25377545
#> 3  -1.42826362 -0.280563081  0.58969393   2     X1  0.80942993
#> 4  -1.42826362 -0.280563081  0.58969393   2     X2 -1.03438352
#> 5  -1.19696192 -0.464708181  1.73893453   3     X1  1.75324463
#> 6  -1.19696192 -0.464708181  1.73893453   3     X2  1.27786205
#> 7  -0.12350154 -0.153740777 -0.47436713   4     X1  0.34033378
#> 8  -0.12350154 -0.153740777 -0.47436713   4     X2 -0.49315289
#> 9  -0.32661929  0.440330269  1.06335275   5     X1  0.05742547
#> 10 -0.32661929  0.440330269  1.06335275   5     X2 -0.15056658
#> 11  0.04835845 -0.819096837 -0.04627308   6     X1 -1.07228728
#> 12  0.04835845 -0.819096837 -0.04627308   6     X2  0.30460086
#> 13 -0.27851777 -0.012241473  0.59851817   7     X1 -1.39368062
#> 14 -0.27851777 -0.012241473  0.59851817   7     X2  0.72508795
#> 15  0.17570853 -0.009096347 -1.78389835   8     X1 -0.47811111
#> 16  0.17570853 -0.009096347 -1.78389835   8     X2 -1.42282656
#> 17  0.87733828 -0.644358171 -0.69071486   9     X1 -0.19723478
#> 18  0.87733828 -0.644358171 -0.69071486   9     X2 -1.78800916
#> 19  1.16033231 -1.832799747 -1.36756016  10     X1 -0.95699731
#> 20  1.16033231 -1.832799747 -1.36756016  10     X2  1.85027785

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID    Value_X1   Value_X2    Value_X3     Value_X4    Value_X5
#> 1       1 -0.53100536 -0.2537754 -1.44435733 -0.655343538  0.43599360
#> 2       2  0.80942993 -1.0343835 -1.42826362 -0.280563081  0.58969393
#> 3       3  1.75324463  1.2778620 -1.19696192 -0.464708181  1.73893453
#> 4       4  0.34033378 -0.4931529 -0.12350154 -0.153740777 -0.47436713
#> 5       5  0.05742547 -0.1505666 -0.32661929  0.440330269  1.06335275
#> 6       6 -1.07228728  0.3046009  0.04835845 -0.819096837 -0.04627308
#> 7       7 -1.39368062  0.7250879 -0.27851777 -0.012241473  0.59851817
#> 8       8 -0.47811111 -1.4228266  0.17570853 -0.009096347 -1.78389835
#> 9       9 -0.19723478 -1.7880092  0.87733828 -0.644358171 -0.69071486
#> 10     10 -0.95699731  1.8502778  1.16033231 -1.832799747 -1.36756016

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5