Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns.

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name       Value
#> 1    X1  1.62093425
#> 2    X2  0.60234481
#> 3    X3 -1.68000342
#> 4    X4  0.22080980
#> 5    X5 -1.74147304
#> 6    X1  0.52983516
#> 7    X2  0.27399370
#> 8    X3  0.02095979
#> 9    X4  0.56508812
#> 10   X5 -1.17554519
#> 11   X1 -0.20311140
#> 12   X2  1.37995272
#> 13   X3 -0.90659210
#> 14   X4  1.42809279
#> 15   X5 -0.17062747
#> 16   X1 -2.90481100
#> 17   X2  0.01530401
#> 18   X3 -1.82504842
#> 19   X4  0.65595749
#> 20   X5 -1.39215132
#> 21   X1  1.05714032
#> 22   X2 -1.24299189
#> 23   X3 -0.58000586
#> 24   X4  0.32832540
#> 25   X5  0.24081944
#> 26   X1 -0.13002913
#> 27   X2 -0.63369036
#> 28   X3  0.83176752
#> 29   X4 -1.30294785
#> 30   X5  1.16261104
#> 31   X1 -0.99274680
#> 32   X2  1.44925850
#> 33   X3 -0.69332808
#> 34   X4 -0.05523798
#> 35   X5  1.12810555
#> 36   X1  0.28313030
#> 37   X2 -0.10268224
#> 38   X3 -0.02889145
#> 39   X4 -0.55082664
#> 40   X5 -1.30037793
#> 41   X1  0.25035578
#> 42   X2 -0.06086260
#> 43   X3 -0.20409064
#> 44   X4  1.30443270
#> 45   X5 -0.44735972
#> 46   X1 -0.76706515
#> 47   X2 -0.12571313
#> 48   X3 -0.42320630
#> 49   X4 -0.29135734
#> 50   X5  0.27123120

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>             X3          X4         X5 Row Column     Numbers
#> 1  -1.68000342  0.22080980 -1.7414730   1     X1  1.62093425
#> 2  -1.68000342  0.22080980 -1.7414730   1     X2  0.60234481
#> 3   0.02095979  0.56508812 -1.1755452   2     X1  0.52983516
#> 4   0.02095979  0.56508812 -1.1755452   2     X2  0.27399370
#> 5  -0.90659210  1.42809279 -0.1706275   3     X1 -0.20311140
#> 6  -0.90659210  1.42809279 -0.1706275   3     X2  1.37995272
#> 7  -1.82504842  0.65595749 -1.3921513   4     X1 -2.90481100
#> 8  -1.82504842  0.65595749 -1.3921513   4     X2  0.01530401
#> 9  -0.58000586  0.32832540  0.2408194   5     X1  1.05714032
#> 10 -0.58000586  0.32832540  0.2408194   5     X2 -1.24299189
#> 11  0.83176752 -1.30294785  1.1626110   6     X1 -0.13002913
#> 12  0.83176752 -1.30294785  1.1626110   6     X2 -0.63369036
#> 13 -0.69332808 -0.05523798  1.1281055   7     X1 -0.99274680
#> 14 -0.69332808 -0.05523798  1.1281055   7     X2  1.44925850
#> 15 -0.02889145 -0.55082664 -1.3003779   8     X1  0.28313030
#> 16 -0.02889145 -0.55082664 -1.3003779   8     X2 -0.10268224
#> 17 -0.20409064  1.30443270 -0.4473597   9     X1  0.25035578
#> 18 -0.20409064  1.30443270 -0.4473597   9     X2 -0.06086260
#> 19 -0.42320630 -0.29135734  0.2712312  10     X1 -0.76706515
#> 20 -0.42320630 -0.29135734  0.2712312  10     X2 -0.12571313

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID   Value_X1    Value_X2    Value_X3    Value_X4   Value_X5
#> 1       1  1.6209343  0.60234481 -1.68000342  0.22080980 -1.7414730
#> 2       2  0.5298352  0.27399370  0.02095979  0.56508812 -1.1755452
#> 3       3 -0.2031114  1.37995272 -0.90659210  1.42809279 -0.1706275
#> 4       4 -2.9048110  0.01530401 -1.82504842  0.65595749 -1.3921513
#> 5       5  1.0571403 -1.24299189 -0.58000586  0.32832540  0.2408194
#> 6       6 -0.1300291 -0.63369036  0.83176752 -1.30294785  1.1626110
#> 7       7 -0.9927468  1.44925850 -0.69332808 -0.05523798  1.1281055
#> 8       8  0.2831303 -0.10268224 -0.02889145 -0.55082664 -1.3003779
#> 9       9  0.2503558 -0.06086260 -0.20409064  1.30443270 -0.4473597
#> 10     10 -0.7670652 -0.12571313 -0.42320630 -0.29135734  0.2712312

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5