Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

Usage

data_to_long(
  data,
  select = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ignore_case = FALSE,
  regex = FALSE,
  cols = select,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  select = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ignore_case = FALSE,
  regex = FALSE,
  cols = select,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

select

Variables that will be included when performing the required tasks. Can be either

a variable specified as a literal variable name (e.g., column_name),

a string with the variable name (e.g., "column_name"), or a character vector of variable names (e.g., c("col1", "col2", "col3")),

a formula with variable names (e.g., ~column_1 + column_2),

a vector of positive integers, giving the positions counting from the left (e.g. 1 or c(1, 3, 5)),

a vector of negative integers, giving the positions counting from the right (e.g., -1 or -1:-3),

or one of the following select-helpers: starts_with(""), ends_with(""), contains(""), a range using : or regex("").

If NULL, selects all columns.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

ignore_case

Logical, if TRUE and when one of the select-helpers or a regular expression is used in select, ignores lower/upper case in the search pattern when matching against variable names.

regex

Logical, if TRUE, the search pattern from select will be treated as regular expression. When regex = TRUE, select must be a character string (or a variable containing a character string) and is not allowed to be one of the supported select-helpers or a character vector of length > 1. regex = TRUE is comparable to using one of the two select-helpers, select = contains("") or select = regex(""), however, since the select-helpers may not work when called from inside other functions (see 'Details'), this argument may be used as workaround.

cols

Deprecated. Please use select.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns.

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

See also

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name       Value
#> 1    X1  0.11068272
#> 2    X2 -0.72889123
#> 3    X3  0.87813349
#> 4    X4  2.16895597
#> 5    X5  1.36860228
#> 6    X1 -0.55584113
#> 7    X2 -0.62503927
#> 8    X3  0.82158108
#> 9    X4  1.20796200
#> 10   X5 -0.22577099
#> 11   X1  1.78691314
#> 12   X2 -1.68669331
#> 13   X3  0.68864025
#> 14   X4 -1.12310858
#> 15   X5  1.51647060
#> 16   X1  0.49785048
#> 17   X2  0.83778704
#> 18   X3  0.55391765
#> 19   X4 -0.40288484
#> 20   X5 -1.54875280
#> 21   X1 -1.96661716
#> 22   X2  0.15337312
#> 23   X3 -0.06191171
#> 24   X4 -0.46665535
#> 25   X5  0.58461375
#> 26   X1  0.70135590
#> 27   X2 -1.13813694
#> 28   X3 -0.30596266
#> 29   X4  0.77996512
#> 30   X5  0.12385424
#> 31   X1 -0.47279141
#> 32   X2  1.25381492
#> 33   X3 -0.38047100
#> 34   X4 -0.08336907
#> 35   X5  0.21594157
#> 36   X1 -1.06782371
#> 37   X2  0.42646422
#> 38   X3 -0.69470698
#> 39   X4  0.25331851
#> 40   X5  0.37963948
#> 41   X1 -0.21797491
#> 42   X2 -0.29507148
#> 43   X3 -0.20791728
#> 44   X4 -0.02854676
#> 45   X5 -0.50232345
#> 46   X1 -1.02600445
#> 47   X2  0.89512566
#> 48   X3 -1.26539635
#> 49   X4 -0.04287046
#> 50   X5 -0.33320738

# Customizing the names
data_to_long(wide_data,
  select = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>             X3          X4         X5 Row Column    Numbers
#> 1   0.87813349  2.16895597  1.3686023   1     X1  0.1106827
#> 2   0.87813349  2.16895597  1.3686023   1     X2 -0.7288912
#> 3   0.82158108  1.20796200 -0.2257710   2     X1 -0.5558411
#> 4   0.82158108  1.20796200 -0.2257710   2     X2 -0.6250393
#> 5   0.68864025 -1.12310858  1.5164706   3     X1  1.7869131
#> 6   0.68864025 -1.12310858  1.5164706   3     X2 -1.6866933
#> 7   0.55391765 -0.40288484 -1.5487528   4     X1  0.4978505
#> 8   0.55391765 -0.40288484 -1.5487528   4     X2  0.8377870
#> 9  -0.06191171 -0.46665535  0.5846137   5     X1 -1.9666172
#> 10 -0.06191171 -0.46665535  0.5846137   5     X2  0.1533731
#> 11 -0.30596266  0.77996512  0.1238542   6     X1  0.7013559
#> 12 -0.30596266  0.77996512  0.1238542   6     X2 -1.1381369
#> 13 -0.38047100 -0.08336907  0.2159416   7     X1 -0.4727914
#> 14 -0.38047100 -0.08336907  0.2159416   7     X2  1.2538149
#> 15 -0.69470698  0.25331851  0.3796395   8     X1 -1.0678237
#> 16 -0.69470698  0.25331851  0.3796395   8     X2  0.4264642
#> 17 -0.20791728 -0.02854676 -0.5023235   9     X1 -0.2179749
#> 18 -0.20791728 -0.02854676 -0.5023235   9     X2 -0.2950715
#> 19 -1.26539635 -0.04287046 -0.3332074  10     X1 -1.0260044
#> 20 -1.26539635 -0.04287046 -0.3332074  10     X2  0.8951257

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID   Value_X1   Value_X2    Value_X3    Value_X4   Value_X5
#> 1       1  0.1106827 -0.7288912  0.87813349  2.16895597  1.3686023
#> 2       2 -0.5558411 -0.6250393  0.82158108  1.20796200 -0.2257710
#> 3       3  1.7869131 -1.6866933  0.68864025 -1.12310858  1.5164706
#> 4       4  0.4978505  0.8377870  0.55391765 -0.40288484 -1.5487528
#> 5       5 -1.9666172  0.1533731 -0.06191171 -0.46665535  0.5846137
#> 6       6  0.7013559 -1.1381369 -0.30596266  0.77996512  0.1238542
#> 7       7 -0.4727914  1.2538149 -0.38047100 -0.08336907  0.2159416
#> 8       8 -1.0678237  0.4264642 -0.69470698  0.25331851  0.3796395
#> 9       9 -0.2179749 -0.2950715 -0.20791728 -0.02854676 -0.5023235
#> 10     10 -1.0260044  0.8951257 -1.26539635 -0.04287046 -0.3332074

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    select = regex("\\d"), # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5