Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns.

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name       Value
#> 1    X1 -0.37121827
#> 2    X2  0.40957481
#> 3    X3 -0.07320776
#> 4    X4  0.19261492
#> 5    X5  0.11203522
#> 6    X1  2.57164649
#> 7    X2  1.55725012
#> 8    X3  1.47815617
#> 9    X4  0.48642298
#> 10   X5  0.07296954
#> 11   X1  0.07891744
#> 12   X2  0.88562601
#> 13   X3 -0.96599602
#> 14   X4  0.26724635
#> 15   X5 -0.95984382
#> 16   X1 -1.77058745
#> 17   X2 -0.23037691
#> 18   X3 -2.52558316
#> 19   X4  0.45784556
#> 20   X5  1.02088718
#> 21   X1  0.33272454
#> 22   X2 -0.12516490
#> 23   X3  0.07812378
#> 24   X4  0.47697968
#> 25   X5 -1.12604256
#> 26   X1  0.34922641
#> 27   X2  0.08880893
#> 28   X3  3.00920649
#> 29   X4  0.04027386
#> 30   X5 -0.36768979
#> 31   X1 -1.05280183
#> 32   X2  1.46356304
#> 33   X3 -1.11601507
#> 34   X4 -0.02167604
#> 35   X5  1.31618932
#> 36   X1 -1.42519899
#> 37   X2 -1.25728162
#> 38   X3  1.67513432
#> 39   X4 -1.04314302
#> 40   X5 -0.84951474
#> 41   X1 -1.22167908
#> 42   X2  1.53961458
#> 43   X3 -2.04888662
#> 44   X4  0.23050000
#> 45   X5  0.75289623
#> 46   X1 -1.46662326
#> 47   X2  1.03356843
#> 48   X3 -1.05430434
#> 49   X4  1.74575038
#> 50   X5  0.06855850

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>             X3          X4          X5 Row Column     Numbers
#> 1  -0.07320776  0.19261492  0.11203522   1     X1 -0.37121827
#> 2  -0.07320776  0.19261492  0.11203522   1     X2  0.40957481
#> 3   1.47815617  0.48642298  0.07296954   2     X1  2.57164649
#> 4   1.47815617  0.48642298  0.07296954   2     X2  1.55725012
#> 5  -0.96599602  0.26724635 -0.95984382   3     X1  0.07891744
#> 6  -0.96599602  0.26724635 -0.95984382   3     X2  0.88562601
#> 7  -2.52558316  0.45784556  1.02088718   4     X1 -1.77058745
#> 8  -2.52558316  0.45784556  1.02088718   4     X2 -0.23037691
#> 9   0.07812378  0.47697968 -1.12604256   5     X1  0.33272454
#> 10  0.07812378  0.47697968 -1.12604256   5     X2 -0.12516490
#> 11  3.00920649  0.04027386 -0.36768979   6     X1  0.34922641
#> 12  3.00920649  0.04027386 -0.36768979   6     X2  0.08880893
#> 13 -1.11601507 -0.02167604  1.31618932   7     X1 -1.05280183
#> 14 -1.11601507 -0.02167604  1.31618932   7     X2  1.46356304
#> 15  1.67513432 -1.04314302 -0.84951474   8     X1 -1.42519899
#> 16  1.67513432 -1.04314302 -0.84951474   8     X2 -1.25728162
#> 17 -2.04888662  0.23050000  0.75289623   9     X1 -1.22167908
#> 18 -2.04888662  0.23050000  0.75289623   9     X2  1.53961458
#> 19 -1.05430434  1.74575038  0.06855850  10     X1 -1.46662326
#> 20 -1.05430434  1.74575038  0.06855850  10     X2  1.03356843

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID    Value_X1    Value_X2    Value_X3    Value_X4    Value_X5
#> 1       1 -0.37121827  0.40957481 -0.07320776  0.19261492  0.11203522
#> 2       2  2.57164649  1.55725012  1.47815617  0.48642298  0.07296954
#> 3       3  0.07891744  0.88562601 -0.96599602  0.26724635 -0.95984382
#> 4       4 -1.77058745 -0.23037691 -2.52558316  0.45784556  1.02088718
#> 5       5  0.33272454 -0.12516490  0.07812378  0.47697968 -1.12604256
#> 6       6  0.34922641  0.08880893  3.00920649  0.04027386 -0.36768979
#> 7       7 -1.05280183  1.46356304 -1.11601507 -0.02167604  1.31618932
#> 8       8 -1.42519899 -1.25728162  1.67513432 -1.04314302 -0.84951474
#> 9       9 -1.22167908  1.53961458 -2.04888662  0.23050000  0.75289623
#> 10     10 -1.46662326  1.03356843 -1.05430434  1.74575038  0.06855850

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5