Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

Usage

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns.

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name       Value
#> 1    X1 -0.81512863
#> 2    X2 -0.22305198
#> 3    X3 -1.74723868
#> 4    X4  1.53865909
#> 5    X5  0.97789743
#> 6    X1  0.16956568
#> 7    X2  0.49127042
#> 8    X3  0.63736377
#> 9    X4  0.22691361
#> 10   X5 -0.87788871
#> 11   X1 -0.40161410
#> 12   X2  0.98291723
#> 13   X3 -1.35924080
#> 14   X4  0.78701577
#> 15   X5 -0.96079976
#> 16   X1 -1.90482886
#> 17   X2  1.15439196
#> 18   X3 -0.31033697
#> 19   X4 -0.16923587
#> 20   X5 -0.54145432
#> 21   X1 -1.06171043
#> 22   X2  1.26056664
#> 23   X3  0.33073222
#> 24   X4 -2.59825155
#> 25   X5 -0.25019220
#> 26   X1 -0.29941959
#> 27   X2 -0.05555054
#> 28   X3 -0.27393011
#> 29   X4  0.75009489
#> 30   X5 -0.66292627
#> 31   X1 -1.00667575
#> 32   X2  2.15218486
#> 33   X3  0.32795180
#> 34   X4 -0.91361432
#> 35   X5 -1.77422148
#> 36   X1 -0.72569594
#> 37   X2  1.61113382
#> 38   X3  1.04053963
#> 39   X4  0.43022208
#> 40   X5  0.03428085
#> 41   X1 -1.00556678
#> 42   X2 -0.59006591
#> 43   X3  0.73481756
#> 44   X4  1.01740672
#> 45   X5  0.97941156
#> 46   X1  1.34528155
#> 47   X2 -0.64829149
#> 48   X3 -0.89559444
#> 49   X4 -0.92914718
#> 50   X5 -0.25725389

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>            X3         X4          X5 Row Column     Numbers
#> 1  -1.7472387  1.5386591  0.97789743   1     X1 -0.81512863
#> 2  -1.7472387  1.5386591  0.97789743   1     X2 -0.22305198
#> 3   0.6373638  0.2269136 -0.87788871   2     X1  0.16956568
#> 4   0.6373638  0.2269136 -0.87788871   2     X2  0.49127042
#> 5  -1.3592408  0.7870158 -0.96079976   3     X1 -0.40161410
#> 6  -1.3592408  0.7870158 -0.96079976   3     X2  0.98291723
#> 7  -0.3103370 -0.1692359 -0.54145432   4     X1 -1.90482886
#> 8  -0.3103370 -0.1692359 -0.54145432   4     X2  1.15439196
#> 9   0.3307322 -2.5982516 -0.25019220   5     X1 -1.06171043
#> 10  0.3307322 -2.5982516 -0.25019220   5     X2  1.26056664
#> 11 -0.2739301  0.7500949 -0.66292627   6     X1 -0.29941959
#> 12 -0.2739301  0.7500949 -0.66292627   6     X2 -0.05555054
#> 13  0.3279518 -0.9136143 -1.77422148   7     X1 -1.00667575
#> 14  0.3279518 -0.9136143 -1.77422148   7     X2  2.15218486
#> 15  1.0405396  0.4302221  0.03428085   8     X1 -0.72569594
#> 16  1.0405396  0.4302221  0.03428085   8     X2  1.61113382
#> 17  0.7348176  1.0174067  0.97941156   9     X1 -1.00556678
#> 18  0.7348176  1.0174067  0.97941156   9     X2 -0.59006591
#> 19 -0.8955944 -0.9291472 -0.25725389  10     X1  1.34528155
#> 20 -0.8955944 -0.9291472 -0.25725389  10     X2 -0.64829149

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID   Value_X1    Value_X2   Value_X3   Value_X4    Value_X5
#> 1       1 -0.8151286 -0.22305198 -1.7472387  1.5386591  0.97789743
#> 2       2  0.1695657  0.49127042  0.6373638  0.2269136 -0.87788871
#> 3       3 -0.4016141  0.98291723 -1.3592408  0.7870158 -0.96079976
#> 4       4 -1.9048289  1.15439196 -0.3103370 -0.1692359 -0.54145432
#> 5       5 -1.0617104  1.26056664  0.3307322 -2.5982516 -0.25019220
#> 6       6 -0.2994196 -0.05555054 -0.2739301  0.7500949 -0.66292627
#> 7       7 -1.0066758  2.15218486  0.3279518 -0.9136143 -1.77422148
#> 8       8 -0.7256959  1.61113382  1.0405396  0.4302221  0.03428085
#> 9       9 -1.0055668 -0.59006591  0.7348176  1.0174067  0.97941156
#> 10     10  1.3452815 -0.64829149 -0.8955944 -0.9291472 -0.25725389

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5