Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

Usage

data_to_long(
  data,
  select = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ignore_case = FALSE,
  regex = FALSE,
  cols = select,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from,
  verbose = TRUE
)

reshape_longer(
  data,
  select = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ignore_case = FALSE,
  regex = FALSE,
  cols = select,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from,
  verbose = TRUE
)

Arguments

data

A data frame to pivot.

select

Variables that will be included when performing the required tasks. Can be either

a variable specified as a literal variable name (e.g., column_name),

a string with the variable name (e.g., "column_name"), or a character vector of variable names (e.g., c("col1", "col2", "col3")),

a formula with variable names (e.g., ~column_1 + column_2),

a vector of positive integers, giving the positions counting from the left (e.g. 1 or c(1, 3, 5)),

a vector of negative integers, giving the positions counting from the right (e.g., -1 or -1:-3),

one of the following select-helpers: starts_with(""), ends_with(""), contains(""), a range using : or regex(""),

or a function testing for logical conditions, e.g. is.numeric() (or is.numeric), or any user-defined function that selects the variables for which the function returns TRUE (like: foo <- function(x) mean(x) > 3),

ranges specified via literal variable names, select-helpers (except regex()) and (user-defined) functions can be negated, i.e. return non-matching elements, when prefixed with a -, e.g. -ends_with(""), -is.numeric or -Sepal.Width:Petal.Length. Note: Negation means that matches are excluded, and thus, the exclude argument can be used alternatively. For instance, select=-ends_with("Length") (with -) is equivalent to exclude=ends_with("Length") (no -). In case negation should not work as expected, use the exclude argument instead.

If NULL, selects all columns. Patterns that found no matches are silently ignored, e.g. find_columns(iris, select = c("Species", "Test")) will just return "Species".

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row names or row numbers from the original data. If NULL, will be removed.

ignore_case

Logical, if TRUE and when one of the select-helpers or a regular expression is used in select, ignores lower/upper case in the search pattern when matching against variable names.

regex

Logical, if TRUE, the search pattern from select will be treated as regular expression. When regex = TRUE, select must be a character string (or a variable containing a character string) and is not allowed to be one of the supported select-helpers or a character vector of length > 1. regex = TRUE is comparable to using one of the two select-helpers, select = contains("") or select = regex(""), however, since the select-helpers may not work when called from inside other functions (see 'Details'), this argument may be used as workaround.

cols

Deprecated. Please use select.

...

Currently not used.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values to be used as future variable values.

colnames_from

The name of the column that contains the levels to be used as future column names.

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

verbose

Toggle warnings.

Value

data.frame

See also

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name       Value
#> 1    X1  0.96438796
#> 2    X2 -0.05906520
#> 3    X3 -0.49945619
#> 4    X4 -1.18164198
#> 5    X5 -0.26924131
#> 6    X1  0.78540975
#> 7    X2 -0.66621750
#> 8    X3  1.11673935
#> 9    X4 -0.43727481
#> 10   X5 -0.28183866
#> 11   X1 -0.68673736
#> 12   X2 -0.83318830
#> 13   X3  0.17915927
#> 14   X4 -0.27775661
#> 15   X5  1.62361443
#> 16   X1 -0.35491405
#> 17   X2 -1.41166020
#> 18   X3 -1.14038587
#> 19   X4 -0.65407266
#> 20   X5  0.82933164
#> 21   X1 -0.25090623
#> 22   X2  0.22522310
#> 23   X3  0.80261795
#> 24   X4  0.18375250
#> 25   X5 -0.79579710
#> 26   X1  0.71122444
#> 27   X2  1.30430074
#> 28   X3 -0.01983587
#> 29   X4 -0.67286229
#> 30   X5 -0.52205496
#> 31   X1  1.69081622
#> 32   X2  0.28000624
#> 33   X3  0.73973690
#> 34   X4  1.98316979
#> 35   X5 -0.02966633
#> 36   X1  0.21996840
#> 37   X2  1.21381633
#> 38   X3 -1.76563045
#> 39   X4  0.98711564
#> 40   X5  0.72787721
#> 41   X1 -0.88773194
#> 42   X2 -0.53360955
#> 43   X3  0.74375283
#> 44   X4 -0.49170812
#> 45   X5 -2.91160978
#> 46   X1 -0.91386940
#> 47   X2  0.78988333
#> 48   X3 -0.44083106
#> 49   X4 -0.35785520
#> 50   X5 -1.48411421

# Customizing the names
data_to_long(wide_data,
  select = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>             X3         X4          X5 Row Column    Numbers
#> 1  -0.49945619 -1.1816420 -0.26924131   1     X1  0.9643880
#> 2  -0.49945619 -1.1816420 -0.26924131   1     X2 -0.0590652
#> 3   1.11673935 -0.4372748 -0.28183866   2     X1  0.7854097
#> 4   1.11673935 -0.4372748 -0.28183866   2     X2 -0.6662175
#> 5   0.17915927 -0.2777566  1.62361443   3     X1 -0.6867374
#> 6   0.17915927 -0.2777566  1.62361443   3     X2 -0.8331883
#> 7  -1.14038587 -0.6540727  0.82933164   4     X1 -0.3549140
#> 8  -1.14038587 -0.6540727  0.82933164   4     X2 -1.4116602
#> 9   0.80261795  0.1837525 -0.79579710   5     X1 -0.2509062
#> 10  0.80261795  0.1837525 -0.79579710   5     X2  0.2252231
#> 11 -0.01983587 -0.6728623 -0.52205496   6     X1  0.7112244
#> 12 -0.01983587 -0.6728623 -0.52205496   6     X2  1.3043007
#> 13  0.73973690  1.9831698 -0.02966633   7     X1  1.6908162
#> 14  0.73973690  1.9831698 -0.02966633   7     X2  0.2800062
#> 15 -1.76563045  0.9871156  0.72787721   8     X1  0.2199684
#> 16 -1.76563045  0.9871156  0.72787721   8     X2  1.2138163
#> 17  0.74375283 -0.4917081 -2.91160978   9     X1 -0.8877319
#> 18  0.74375283 -0.4917081 -2.91160978   9     X2 -0.5336095
#> 19 -0.44083106 -0.3578552 -1.48411421  10     X1 -0.9138694
#> 20 -0.44083106 -0.3578552 -1.48411421  10     X2  0.7898833

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID         X1         X2          X3         X4          X5
#> 1       1  0.9643880 -0.0590652 -0.49945619 -1.1816420 -0.26924131
#> 2       2  0.7854097 -0.6662175  1.11673935 -0.4372748 -0.28183866
#> 3       3 -0.6867374 -0.8331883  0.17915927 -0.2777566  1.62361443
#> 4       4 -0.3549140 -1.4116602 -1.14038587 -0.6540727  0.82933164
#> 5       5 -0.2509062  0.2252231  0.80261795  0.1837525 -0.79579710
#> 6       6  0.7112244  1.3043007 -0.01983587 -0.6728623 -0.52205496
#> 7       7  1.6908162  0.2800062  0.73973690  1.9831698 -0.02966633
#> 8       8  0.2199684  1.2138163 -1.76563045  0.9871156  0.72787721
#> 9       9 -0.8877319 -0.5336095  0.74375283 -0.4917081 -2.91160978
#> 10     10 -0.9138694  0.7898833 -0.44083106 -0.3578552 -1.48411421

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    select = regex("\\d"), # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet I1 I2 I3 I4 I5
#> 1      1        NA  16       61617     A  2  4  3  4  4
#> 2      1        NA  16       61617     C  2  3  3  4  4
#> 3      1        NA  16       61617     E  3  3  3  4  4
#> 4      1        NA  16       61617     N  3  4  2  2  3
#> 5      1        NA  16       61617     O  3  6  3  4  3
#> 6      2        NA  18       61618     A  2  4  5  2  5