Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

Usage

data_to_long(
  data,
  select = "all",
  names_to = "name",
  names_prefix = NULL,
  names_sep = NULL,
  names_pattern = NULL,
  values_to = "value",
  values_drop_na = FALSE,
  rows_to = NULL,
  ignore_case = FALSE,
  regex = FALSE,
  ...,
  cols,
  colnames_to
)

reshape_longer(
  data,
  select = "all",
  names_to = "name",
  names_prefix = NULL,
  names_sep = NULL,
  names_pattern = NULL,
  values_to = "value",
  values_drop_na = FALSE,
  rows_to = NULL,
  ignore_case = FALSE,
  regex = FALSE,
  ...,
  cols,
  colnames_to
)

Arguments

data

A data frame to pivot.

select

Variables that will be included when performing the required tasks. Can be either

a variable specified as a literal variable name (e.g., column_name),

a string with the variable name (e.g., "column_name"), or a character vector of variable names (e.g., c("col1", "col2", "col3")),

a formula with variable names (e.g., ~column_1 + column_2),

a vector of positive integers, giving the positions counting from the left (e.g. 1 or c(1, 3, 5)),

a vector of negative integers, giving the positions counting from the right (e.g., -1 or -1:-3),

one of the following select-helpers: starts_with(), ends_with(), contains(), a range using : or regex(""). starts_with(), ends_with(), and contains() accept several patterns, e.g starts_with("Sep", "Petal").

or a function testing for logical conditions, e.g. is.numeric() (or is.numeric), or any user-defined function that selects the variables for which the function returns TRUE (like: foo <- function(x) mean(x) > 3),

ranges specified via literal variable names, select-helpers (except regex()) and (user-defined) functions can be negated, i.e. return non-matching elements, when prefixed with a -, e.g. -ends_with(""), -is.numeric or -Sepal.Width:Petal.Length. Note: Negation means that matches are excluded, and thus, the exclude argument can be used alternatively. For instance, select=-ends_with("Length") (with -) is equivalent to exclude=ends_with("Length") (no -). In case negation should not work as expected, use the exclude argument instead.

If NULL, selects all columns. Patterns that found no matches are silently ignored, e.g. find_columns(iris, select = c("Species", "Test")) will just return "Species".

names_to

The name of the new column that will contain the column names.

names_prefix

A regular expression used to remove matching text from the start of each variable name.

names_sep, names_pattern

If names_to contains multiple values, this argument controls how the column name is broken up. names_pattern takes a regular expression containing matching groups, i.e. "()".

values_to

The name of the new column that will contain the values of the pivoted variables.

values_drop_na

If TRUE, will drop rows that contain only NA in the values_to column. This effectively converts explicit missing values to implicit missing values, and should generally be used only when missing values in data were created by its structure.

rows_to

The name of the column that will contain the row names or row numbers from the original data. If NULL, will be removed.

ignore_case

Logical, if TRUE and when one of the select-helpers or a regular expression is used in select, ignores lower/upper case in the search pattern when matching against variable names.

regex

Logical, if TRUE, the search pattern from select will be treated as regular expression. When regex = TRUE, select must be a character string (or a variable containing a character string) and is not allowed to be one of the supported select-helpers or a character vector of length > 1. regex = TRUE is comparable to using one of the two select-helpers, select = contains("") or select = regex(""), however, since the select-helpers may not work when called from inside other functions (see 'Details'), this argument may be used as workaround.

...

Currently not used.

cols

Identical to select. This argument is here to ensure compatibility with tidyr::pivot_longer(). If both select and cols are provided, cols is used.

colnames_to

Deprecated. Use names_to instead.

Value

If a tibble was provided as input, reshape_longer() also returns a tibble. Otherwise, it returns a data frame.

See also

Examples

# \donttest{
wide_data <- data.frame(replicate(5, rnorm(10)))

# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    name       value
#> 1    X1  0.96438796
#> 2    X2 -0.05906520
#> 3    X3 -0.49945619
#> 4    X4 -1.18164198
#> 5    X5 -0.26924131
#> 6    X1  0.78540975
#> 7    X2 -0.66621750
#> 8    X3  1.11673935
#> 9    X4 -0.43727481
#> 10   X5 -0.28183866
#> 11   X1 -0.68673736
#> 12   X2 -0.83318830
#> 13   X3  0.17915927
#> 14   X4 -0.27775661
#> 15   X5  1.62361443
#> 16   X1 -0.35491405
#> 17   X2 -1.41166020
#> 18   X3 -1.14038587
#> 19   X4 -0.65407266
#> 20   X5  0.82933164
#> 21   X1 -0.25090623
#> 22   X2  0.22522310
#> 23   X3  0.80261795
#> 24   X4  0.18375250
#> 25   X5 -0.79579710
#> 26   X1  0.71122444
#> 27   X2  1.30430074
#> 28   X3 -0.01983587
#> 29   X4 -0.67286229
#> 30   X5 -0.52205496
#> 31   X1  1.69081622
#> 32   X2  0.28000624
#> 33   X3  0.73973690
#> 34   X4  1.98316979
#> 35   X5 -0.02966633
#> 36   X1  0.21996840
#> 37   X2  1.21381633
#> 38   X3 -1.76563045
#> 39   X4  0.98711564
#> 40   X5  0.72787721
#> 41   X1 -0.88773194
#> 42   X2 -0.53360955
#> 43   X3  0.74375283
#> 44   X4 -0.49170812
#> 45   X5 -2.91160978
#> 46   X1 -0.91386940
#> 47   X2  0.78988333
#> 48   X3 -0.44083106
#> 49   X4 -0.35785520
#> 50   X5 -1.48411421

# Customizing the names
data_to_long(wide_data,
  select = c(1, 2),
  names_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>             X3         X4          X5 Row Column    Numbers
#> 1  -0.49945619 -1.1816420 -0.26924131   1     X1  0.9643880
#> 2  -0.49945619 -1.1816420 -0.26924131   1     X2 -0.0590652
#> 3   1.11673935 -0.4372748 -0.28183866   2     X1  0.7854097
#> 4   1.11673935 -0.4372748 -0.28183866   2     X2 -0.6662175
#> 5   0.17915927 -0.2777566  1.62361443   3     X1 -0.6867374
#> 6   0.17915927 -0.2777566  1.62361443   3     X2 -0.8331883
#> 7  -1.14038587 -0.6540727  0.82933164   4     X1 -0.3549140
#> 8  -1.14038587 -0.6540727  0.82933164   4     X2 -1.4116602
#> 9   0.80261795  0.1837525 -0.79579710   5     X1 -0.2509062
#> 10  0.80261795  0.1837525 -0.79579710   5     X2  0.2252231
#> 11 -0.01983587 -0.6728623 -0.52205496   6     X1  0.7112244
#> 12 -0.01983587 -0.6728623 -0.52205496   6     X2  1.3043007
#> 13  0.73973690  1.9831698 -0.02966633   7     X1  1.6908162
#> 14  0.73973690  1.9831698 -0.02966633   7     X2  0.2800062
#> 15 -1.76563045  0.9871156  0.72787721   8     X1  0.2199684
#> 16 -1.76563045  0.9871156  0.72787721   8     X2  1.2138163
#> 17  0.74375283 -0.4917081 -2.91160978   9     X1 -0.8877319
#> 18  0.74375283 -0.4917081 -2.91160978   9     X2 -0.5336095
#> 19 -0.44083106 -0.3578552 -1.48411421  10     X1 -0.9138694
#> 20 -0.44083106 -0.3578552 -1.48411421  10     X2  0.7898833

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  data_to_long(data,
    select = regex("\\d"), # Select all columns that contain a digit
    names_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  if (require("tidyr")) {
    reshape_longer(
      tidyr::who,
      select = new_sp_m014:newrel_f65,
      names_to = c("diagnosis", "gender", "age"),
      names_pattern = "new_?(.*)_(.)(.*)",
      values_to = "count"
    )
  }
}
#> Loading required package: psych
#> 
#> Attaching package: ‘psych’
#> The following object is masked from ‘package:datawizard’:
#> 
#>     rescale
#> Loading required package: tidyr
#> 
#> Attaching package: ‘tidyr’
#> The following objects are masked from ‘package:poorman’:
#> 
#>     %>%, all_of, any_of, contains, ends_with, everything, last_col,
#>     matches, num_range, pivot_longer, pivot_wider, replace_na,
#>     starts_with, unite
#> # A tibble: 405,440 × 8
#>    country     iso2  iso3   year diagnosis gender age   count
#>  * <chr>       <chr> <chr> <int> <chr>     <chr>  <chr> <int>
#>  1 Afghanistan AF    AFG    1980 sp        m      014      NA
#>  2 Afghanistan AF    AFG    1980 sp        m      1524     NA
#>  3 Afghanistan AF    AFG    1980 sp        m      2534     NA
#>  4 Afghanistan AF    AFG    1980 sp        m      3544     NA
#>  5 Afghanistan AF    AFG    1980 sp        m      4554     NA
#>  6 Afghanistan AF    AFG    1980 sp        m      5564     NA
#>  7 Afghanistan AF    AFG    1980 sp        m      65       NA
#>  8 Afghanistan AF    AFG    1980 sp        f      014      NA
#>  9 Afghanistan AF    AFG    1980 sp        f      1524     NA
#> 10 Afghanistan AF    AFG    1980 sp        f      2534     NA
#> # … with 405,430 more rows
# }