Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns.

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name       Value
#> 1    X1 -0.70956368
#> 2    X2 -0.10806322
#> 3    X3  0.04131932
#> 4    X4 -0.21850318
#> 5    X5  0.56509638
#> 6    X1 -0.57052364
#> 7    X2  0.65588073
#> 8    X3 -0.07923834
#> 9    X4  0.64181274
#> 10   X5  0.92018947
#> 11   X1 -0.04934568
#> 12   X2 -0.31150010
#> 13   X3  0.15559079
#> 14   X4 -0.21609173
#> 15   X5  0.25186472
#> 16   X1 -0.86282475
#> 17   X2 -0.32528069
#> 18   X3  0.28569314
#> 19   X4  0.48584197
#> 20   X5 -0.65064730
#> 21   X1 -0.39250842
#> 22   X2 -0.76115645
#> 23   X3 -1.44906141
#> 24   X4  0.05363588
#> 25   X5  0.35971539
#> 26   X1  1.50130036
#> 27   X2 -0.42274733
#> 28   X3  0.29899427
#> 29   X4  1.13874478
#> 30   X5  1.58014538
#> 31   X1  0.31904494
#> 32   X2 -0.03365051
#> 33   X3  1.91070182
#> 34   X4  0.73147848
#> 35   X5  1.14086149
#> 36   X1 -1.31071023
#> 37   X2 -0.03572529
#> 38   X3  1.06982688
#> 39   X4  0.12365569
#> 40   X5  0.11458967
#> 41   X1  0.07727690
#> 42   X2 -2.08459201
#> 43   X3  0.08708101
#> 44   X4 -0.85023713
#> 45   X5  0.66422586
#> 46   X1  0.99687488
#> 47   X2  0.01943559
#> 48   X3  0.33415109
#> 49   X4 -2.00017347
#> 50   X5 -0.15030896

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>             X3          X4         X5 Row Column     Numbers
#> 1   0.04131932 -0.21850318  0.5650964   1     X1 -0.70956368
#> 2   0.04131932 -0.21850318  0.5650964   1     X2 -0.10806322
#> 3  -0.07923834  0.64181274  0.9201895   2     X1 -0.57052364
#> 4  -0.07923834  0.64181274  0.9201895   2     X2  0.65588073
#> 5   0.15559079 -0.21609173  0.2518647   3     X1 -0.04934568
#> 6   0.15559079 -0.21609173  0.2518647   3     X2 -0.31150010
#> 7   0.28569314  0.48584197 -0.6506473   4     X1 -0.86282475
#> 8   0.28569314  0.48584197 -0.6506473   4     X2 -0.32528069
#> 9  -1.44906141  0.05363588  0.3597154   5     X1 -0.39250842
#> 10 -1.44906141  0.05363588  0.3597154   5     X2 -0.76115645
#> 11  0.29899427  1.13874478  1.5801454   6     X1  1.50130036
#> 12  0.29899427  1.13874478  1.5801454   6     X2 -0.42274733
#> 13  1.91070182  0.73147848  1.1408615   7     X1  0.31904494
#> 14  1.91070182  0.73147848  1.1408615   7     X2 -0.03365051
#> 15  1.06982688  0.12365569  0.1145897   8     X1 -1.31071023
#> 16  1.06982688  0.12365569  0.1145897   8     X2 -0.03572529
#> 17  0.08708101 -0.85023713  0.6642259   9     X1  0.07727690
#> 18  0.08708101 -0.85023713  0.6642259   9     X2 -2.08459201
#> 19  0.33415109 -2.00017347 -0.1503090  10     X1  0.99687488
#> 20  0.33415109 -2.00017347 -0.1503090  10     X2  0.01943559

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID    Value_X1    Value_X2    Value_X3    Value_X4   Value_X5
#> 1       1 -0.70956368 -0.10806322  0.04131932 -0.21850318  0.5650964
#> 2       2 -0.57052364  0.65588073 -0.07923834  0.64181274  0.9201895
#> 3       3 -0.04934568 -0.31150010  0.15559079 -0.21609173  0.2518647
#> 4       4 -0.86282475 -0.32528069  0.28569314  0.48584197 -0.6506473
#> 5       5 -0.39250842 -0.76115645 -1.44906141  0.05363588  0.3597154
#> 6       6  1.50130036 -0.42274733  0.29899427  1.13874478  1.5801454
#> 7       7  0.31904494 -0.03365051  1.91070182  0.73147848  1.1408615
#> 8       8 -1.31071023 -0.03572529  1.06982688  0.12365569  0.1145897
#> 9       9  0.07727690 -2.08459201  0.08708101 -0.85023713  0.6642259
#> 10     10  0.99687488  0.01943559  0.33415109 -2.00017347 -0.1503090

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5