Reshape (pivot) data from long to wide

This function "widens" data, increasing the number of columns and decreasing the number of rows. This is a dependency-free base-R equivalent of tidyr::pivot_wider().

Usage

data_to_wide(
  data,
  id_cols = NULL,
  values_from = "Value",
  names_from = "Name",
  names_sep = "_",
  names_prefix = "",
  names_glue = NULL,
  values_fill = NULL,
  verbose = TRUE,
  ...,
  colnames_from,
  rows_from,
  sep
)

reshape_wider(
  data,
  id_cols = NULL,
  values_from = "Value",
  names_from = "Name",
  names_sep = "_",
  names_prefix = "",
  names_glue = NULL,
  values_fill = NULL,
  verbose = TRUE,
  ...,
  colnames_from,
  rows_from,
  sep
)

Arguments

data

A data frame to pivot.

id_cols

The name of the column that identifies the rows. If NULL, it will use all the unique rows.

values_from

The name of the column that contains the values to be used as future variable values.

names_from

The name of the column that contains the levels to be used as future column names.

names_sep

If names_from or values_from contains multiple variables, this will be used to join their values together into a single string to use as a column name.

names_prefix

String added to the start of every variable name. This is particularly useful if names_from is a numeric vector and you want to create syntactic variable names.

names_glue

Instead of names_sep and names_prefix, you can supply a glue specification that uses the names_from columns to create custom column names. Note that the only delimiters supported by names_glue are curly brackets, { and }.

values_fill

Optionally, a (scalar) value that will be used to replace missing values in the new columns created.

verbose

Toggle warnings.

...

Not used for now.

colnames_from

Deprecated. Use names_from instead.

rows_from

Deprecated. Use id_cols instead.

sep

Deprecated. Use names_sep instead.

Value

If a tibble was provided as input, reshape_wider() also returns a tibble. Otherwise, it returns a data frame.

See also

Examples

data_long <- read.table(header = TRUE, text = "
 subject sex condition measurement
       1   M   control         7.9
       1   M     cond1        12.3
       1   M     cond2        10.7
       2   F   control         6.3
       2   F     cond1        10.6
       2   F     cond2        11.1
       3   F   control         9.5
       3   F     cond1        13.1
       3   F     cond2        13.8
       4   M   control        11.5
       4   M     cond1        13.4
       4   M     cond2        12.9")

data_to_wide(
  data_long,
  id_cols = "subject",
  names_from = "condition",
  values_from = "measurement"
)
#>   subject control cond1 cond2
#> 1       1     7.9  12.3  10.7
#> 2       2     6.3  10.6  11.1
#> 3       3     9.5  13.1  13.8
#> 4       4    11.5  13.4  12.9

data_to_wide(
  data_long,
  id_cols = "subject",
  names_from = "condition",
  values_from = "measurement",
  names_prefix = "Var.",
  names_sep = "."
)
#>   subject Var.control Var.cond1 Var.cond2
#> 1       1         7.9      12.3      10.7
#> 2       2         6.3      10.6      11.1
#> 3       3         9.5      13.1      13.8
#> 4       4        11.5      13.4      12.9

production <- expand.grid(
  product = c("A", "B"),
  country = c("AI", "EI"),
  year = 2000:2014
)
production <- data_filter(production, (product == "A" & country == "AI") | product == "B")

production$production <- rnorm(nrow(production))

data_to_wide(
  production,
  names_from = c("product", "country"),
  values_from = "production",
  names_glue = "prod_{product}_{country}"
)
#>    year  prod_A_AI    prod_B_AI  prod_B_EI
#> 1  2000 -0.8408539  1.430252916  0.3920247
#> 2  2001 -0.4726417 -0.996105337 -0.1950098
#> 3  2002  1.3394131 -0.711765324 -0.9245581
#> 4  2003  0.8737440 -1.043327370  1.0166035
#> 5  2004 -2.2241873  1.878421273 -0.5218175
#> 6  2005 -0.6546695  0.993425211 -0.2819180
#> 7  2006 -1.0952392  1.164258300  0.2246749
#> 8  2007 -1.1649528  0.748724154 -1.3051249
#> 9  2008 -0.3766038  0.004485138  1.5616184
#> 10 2009 -0.7426178 -0.331893557  0.1463996
#> 11 2010  0.4176823  0.036978385 -1.7524488
#> 12 2011  0.1575659  0.411082845 -0.9077312
#> 13 2012  1.4151629 -0.205867410 -0.8926030
#> 14 2013  0.5674379  0.764974595 -1.9997762
#> 15 2014 -1.5185747  0.560874533 -0.8971569